Control.Concurrent.QSemN

Description

Quantity semaphores in which each thread may wait for an arbitrary "amount".

General Quantity Semaphores

data QSemN

QSemN is a quantity semaphore in which the resource is acquired and released in units of one. It provides guaranteed FIFO ordering for satisfying blocked waitQSemN calls.

The pattern

  bracket_ (waitQSemN n) (signalQSemN n) (...)

is safe; it never loses any of the resource.

newQSemN :: Int -> IO QSemN

Build a new QSemN with a supplied initial quantity. The initial quantity must be at least 0.

waitQSemN :: QSemN -> Int -> IO ()

Wait for the specified quantity to become available

signalQSemN :: QSemN -> Int -> IO ()

Signal that a given quantity is now available from the QSemN.