Control.Exception.Context

Description

Exception context and annotations.

Documentation

data ExceptionContext

Exception context represents a list of ExceptionAnnotations. These are attached to SomeExceptions via addExceptionContext and can be used to capture various ad-hoc metadata about the exception including backtraces and application-specific context.

ExceptionContexts can be merged via concatenation using the Semigroup instance or mergeExceptionContext.

Note that GHC will automatically solve implicit constraints of type ExceptionContext with emptyExceptionContext.

Constructors

ExceptionContext [SomeExceptionAnnotation]

emptyExceptionContext :: ExceptionContext

An ExceptionContext containing no annotations.

Since: base-4.20.0.0

addExceptionAnnotation :: ExceptionAnnotation a => a -> ExceptionContext -> ExceptionContext

Construct a singleton ExceptionContext from an ExceptionAnnotation.

Since: base-4.20.0.0

Destructuring

getExceptionAnnotations :: ExceptionAnnotation a => ExceptionContext -> [a]

Retrieve all ExceptionAnnotations of the given type from an ExceptionContext.

Since: base-4.20.0.0

getAllExceptionAnnotations :: ExceptionContext -> [SomeExceptionAnnotation]

displayExceptionContext :: ExceptionContext -> String

Render ExceptionContext to a human-readable String.

Since: base-4.20.0.0