Control.Monad.ST.Lazy.Unsafe

Description

This module presents an identical interface to Control.Monad.ST, except that the monad delays evaluation of ST operations until a value depending on them is required.

Unsafe API.

Unsafe operations

unsafeInterleaveST :: ST s a -> ST s a

unsafeIOToST :: IO a -> ST s a