Data.Complex

Description

Complex numbers.

Rectangular form

data Complex a

A data type representing complex numbers.

You can read about complex numbers on wikipedia.

In haskell, complex numbers are represented as a :+ b which can be thought of as representing \(a + bi\). For a complex number z, abs z is a number with the magnitude of z, but oriented in the positive real direction, whereas signum z has the phase of z, but unit magnitude. Apart from the loss of precision due to IEEE754 floating point numbers, it holds that z == abs z * signum z.

Note that Complex's instances inherit the deficiencies from the type parameter's. For example, Complex Float's Ord instance has similar problems to Float's.

As can be seen in the examples, the Foldable and Traversable instances traverse the real part first.

Examples

Expand

>>> (5.0 :+ 2.5) + 6.5
11.5 :+ 2.5

>>> abs (1.0 :+ 1.0) - sqrt 2.0
0.0 :+ 0.0

>>> abs (signum (4.0 :+ 3.0))
1.0 :+ 0.0

>>> foldr (:) [] (1 :+ 2)
[1,2]

>>> mapM print (1 :+ 2)
1
2

Constructors

!a :+ !a infix 6
forms a complex number from its real and imaginary rectangular components.

realPart :: Complex a -> a

Extracts the real part of a complex number.

Examples

Expand

>>> realPart (5.0 :+ 3.0)
5.0

>>> realPart ((5.0 :+ 3.0) * (2.0 :+ 3.0))
1.0

imagPart :: Complex a -> a

Extracts the imaginary part of a complex number.

Examples

Expand

>>> imagPart (5.0 :+ 3.0)
3.0

>>> imagPart ((5.0 :+ 3.0) * (2.0 :+ 3.0))
21.0

Polar form

mkPolar :: Floating a => a -> a -> Complex a

Form a complex number from polar components of magnitude and phase.

Examples

Expand

>>> mkPolar 1 (pi / 4)
0.7071067811865476 :+ 0.7071067811865475

>>> mkPolar 1 0
1.0 :+ 0.0

cis :: Floating a => a -> Complex a

cis t is a complex value with magnitude 1 and phase t (modulo 2*pi).

cis = mkPolar 1

Examples

Expand

>>> cis 0
1.0 :+ 0.0

The following examples are not perfectly zero due to IEEE 754

>>> cis pi
(-1.0) :+ 1.2246467991473532e-16

>>> cis (4 * pi) - cis (2 * pi)
0.0 :+ (-2.4492935982947064e-16)

polar :: RealFloat a => Complex a -> (a, a)

The function polar takes a complex number and returns a (magnitude, phase) pair in canonical form: the magnitude is non-negative, and the phase in the range (-pi, pi]; if the magnitude is zero, then so is the phase.

polar z = (magnitude z, phase z)

Examples

Expand

>>> polar (1.0 :+ 1.0)
(1.4142135623730951,0.7853981633974483)

>>> polar ((-1.0) :+ 0.0)
(1.0,3.141592653589793)

>>> polar (0.0 :+ 0.0)
(0.0,0.0)

magnitude :: RealFloat a => Complex a -> a

The non-negative magnitude of a complex number.

Examples

Expand

>>> magnitude (1.0 :+ 1.0)
1.4142135623730951

>>> magnitude (1.0 + 0.0)
1.0

>>> magnitude (0.0 :+ (-5.0))
5.0

phase :: RealFloat a => Complex a -> a

The phase of a complex number, in the range (-pi, pi]. If the magnitude is zero, then so is the phase.

Examples

Expand

>>> phase (0.5 :+ 0.5) / pi
0.25

>>> phase (0 :+ 4) / pi
0.5

Conjugate

conjugate :: Num a => Complex a -> Complex a

The conjugate of a complex number.

conjugate (conjugate x) = x

Examples

Expand

>>> conjugate (3.0 :+ 3.0)
3.0 :+ (-3.0)

>>> conjugate ((3.0 :+ 3.0) * (2.0 :+ 2.0))
0.0 :+ (-12.0)