Data.Dynamic

Description

The Dynamic interface provides basic support for dynamic types.

Operations for injecting values of arbitrary type into a dynamically typed value, Dynamic, are provided, together with operations for converting dynamic values into a concrete (monomorphic) type.

The Dynamic type

data Dynamic where

A value of type Dynamic is an object encapsulated together with its type.

A Dynamic may only represent a monomorphic value; an attempt to create a value of type Dynamic from a polymorphically-typed expression will result in an ambiguity error (see toDyn).

Showing a value of type Dynamic returns a pretty-printed representation of the object's type; useful for debugging.

Constructors

Dynamic :: forall a. TypeRep a -> a -> Dynamic

Converting to and from Dynamic

toDyn :: Typeable a => a -> Dynamic

Converts an arbitrary value into an object of type Dynamic.

The type of the object must be an instance of Typeable, which ensures that only monomorphically-typed objects may be converted to Dynamic. To convert a polymorphic object into Dynamic, give it a monomorphic type signature. For example:

   toDyn (id :: Int -> Int)

fromDyn

Arguments

:: Typeable a

=> Dynamic
the dynamically-typed object

-> a
a default value

-> a
returns: the value of the first argument, if it has the correct type, otherwise the value of the second argument.

Converts a Dynamic object back into an ordinary Haskell value of the correct type. See also fromDynamic.

fromDynamic

Arguments

:: forall a. Typeable a

=> Dynamic
the dynamically-typed object

-> Maybe a
returns: Just a, if the dynamically-typed object has the correct type (and a is its value), or Nothing otherwise.

Converts a Dynamic object back into an ordinary Haskell value of the correct type. See also fromDyn.

Applying functions of dynamic type

dynApply :: Dynamic -> Dynamic -> Maybe Dynamic

dynApp :: Dynamic -> Dynamic -> Dynamic

dynTypeRep :: Dynamic -> SomeTypeRep

Convenience re-exports

class Typeable (a :: k)

The class Typeable allows a concrete representation of a type to be calculated.

Minimal complete definition

typeRep#