Data.Functor.Compose

Description

Composition of functors.

Since: base-4.9.0.0

Documentation

newtype Compose (f :: k -> Type) (g :: k1 -> k) (a :: k1) infixr 9

Right-to-left composition of functors. The composition of applicative functors is always applicative, but the composition of monads is not always a monad.

Examples

Expand

>>> fmap (subtract 1) (Compose (Just [1, 2, 3]))
Compose (Just [0,1,2])

>>> Compose (Just [1, 2, 3]) <> Compose Nothing
Compose (Just [1,2,3])

>>> Compose (Just [(++ "World"), (++ "Haskell")]) <*> Compose (Just ["Hello, "])
Compose (Just ["Hello, World","Hello, Haskell"])

Constructors

Compose infixr 9

Fields

getCompose :: f (g a)