Data.Kind

Description

Basic kinds

Since: base-4.9.0.0

Documentation

type Type = TYPE LiftedRep

The kind of types with lifted values. For example Int :: Type.

type Constraint = CONSTRAINT LiftedRep

The kind of lifted constraints

data FUN

The builtin function type, written in infix form as a % m -> b. Values of this type are functions taking inputs of type a and producing outputs of type b. The multiplicity of the input is m.

Note that FUN m a b permits representation polymorphism in both a and b, so that types like Int# -> Int# can still be well-kinded.