Data.OldTypeable.Internal

Description

Deprecated: Use Data.Typeable.Internal instead

The representations of the types TyCon and TypeRep, and the function mkTyCon which is used by derived instances of Typeable to construct a TyCon.

Since: 4.7.0.0

Documentation

data TypeRep

A concrete representation of a (monomorphic) type. TypeRep supports reasonably efficient equality.

Constructors

TypeRep !Fingerprint TyCon [TypeRep]

data TyCon

An abstract representation of a type constructor. TyCon objects can be built using mkTyCon.

Constructors

TyCon

Fields

tyConHash :: !Fingerprint

tyConPackage :: String

tyConModule :: String

tyConName :: String

mkTyCon :: Word# -> Word# -> String -> String -> String -> TyCon

mkTyCon3

Arguments

:: String
package name

-> String
module name

-> String
the name of the type constructor

-> TyCon
A unique TyCon object

Builds a TyCon object representing a type constructor. An implementation of Data.Typeable should ensure that the following holds:

 A==A' ^ B==B' ^ C==C' ==> mkTyCon A B C == mkTyCon A' B' C'

mkTyConApp :: TyCon -> [TypeRep] -> TypeRep

Applies a type constructor to a sequence of types

mkAppTy :: TypeRep -> TypeRep -> TypeRep

Adds a TypeRep argument to a TypeRep.

typeRepTyCon :: TypeRep -> TyCon

Observe the type constructor of a type representation

typeOfDefault :: forall t a. (Typeable1 t, Typeable a) => t a -> TypeRep

For defining a Typeable instance from any Typeable1 instance.

typeOf1Default :: forall t a b. (Typeable2 t, Typeable a) => t a b -> TypeRep

For defining a Typeable1 instance from any Typeable2 instance.

typeOf2Default :: forall t a b c. (Typeable3 t, Typeable a) => t a b c -> TypeRep

For defining a Typeable2 instance from any Typeable3 instance.

typeOf3Default :: forall t a b c d. (Typeable4 t, Typeable a) => t a b c d -> TypeRep

For defining a Typeable3 instance from any Typeable4 instance.

typeOf4Default :: forall t a b c d e. (Typeable5 t, Typeable a) => t a b c d e -> TypeRep

For defining a Typeable4 instance from any Typeable5 instance.

typeOf5Default :: forall t a b c d e f. (Typeable6 t, Typeable a) => t a b c d e f -> TypeRep

For defining a Typeable5 instance from any Typeable6 instance.

typeOf6Default :: forall t a b c d e f g. (Typeable7 t, Typeable a) => t a b c d e f g -> TypeRep

For defining a Typeable6 instance from any Typeable7 instance.

class Typeable a where

The class Typeable allows a concrete representation of a type to be calculated.

Methods

typeOf :: a -> TypeRep

Takes a value of type a and returns a concrete representation of that type. The value of the argument should be ignored by any instance of Typeable, so that it is safe to pass undefined as the argument.

class Typeable1 t where

Variant for unary type constructors

Methods

typeOf1 :: t a -> TypeRep

class Typeable2 t where

Variant for binary type constructors

Methods

typeOf2 :: t a b -> TypeRep

class Typeable3 t where

Variant for 3-ary type constructors

Methods

typeOf3 :: t a b c -> TypeRep

class Typeable4 t where

Variant for 4-ary type constructors

Methods

typeOf4 :: t a b c d -> TypeRep

class Typeable5 t where

Variant for 5-ary type constructors

Methods

typeOf5 :: t a b c d e -> TypeRep

class Typeable6 t where

Variant for 6-ary type constructors

Methods

typeOf6 :: t a b c d e f -> TypeRep

class Typeable7 t where

Variant for 7-ary type constructors

Methods

typeOf7 :: t a b c d e f g -> TypeRep

mkFunTy :: TypeRep -> TypeRep -> TypeRep

A special case of mkTyConApp, which applies the function type constructor to a pair of types.

splitTyConApp :: TypeRep -> (TyCon, [TypeRep])

Splits a type constructor application

funResultTy :: TypeRep -> TypeRep -> Maybe TypeRep

Applies a type to a function type. Returns: Just u if the first argument represents a function of type t -> u and the second argument represents a function of type t. Otherwise, returns Nothing.

typeRepArgs :: TypeRep -> [TypeRep]

Observe the argument types of a type representation

showsTypeRep :: TypeRep -> ShowS

tyConString :: TyCon -> String

Observe string encoding of a type representation

listTc :: TyCon

funTc :: TyCon