Data.Tuple

Description

Functions associated with the tuple data types.

Documentation

fst :: (a, b) -> a

Extract the first component of a pair.

snd :: (a, b) -> b

Extract the second component of a pair.

curry :: ((a, b) -> c) -> a -> b -> c

curry converts an uncurried function to a curried function.

Examples

Expand

>>> curry fst 1 2
1

uncurry :: (a -> b -> c) -> (a, b) -> c

uncurry converts a curried function to a function on pairs.

Examples

Expand

>>> uncurry (+) (1,2)
3

>>> uncurry ($) (show, 1)
"1"

>>> map (uncurry max) [(1,2), (3,4), (6,8)]
[2,4,8]

swap :: (a, b) -> (b, a)

Swap the components of a pair.

data ()

The unit datatype () has one non-undefined member, the nullary constructor ().

Constructors

()

data Unit a

Constructors

Unit a

data (,) a b

Constructors

(,) a b

data (,,) a b c

Constructors

(,,) a b c

data (,,,) a b c d

Constructors

(,,,) a b c d

data (,,,,) a b c d e

Constructors

(,,,,) a b c d e

data (,,,,,) a b c d e f

Constructors

(,,,,,) a b c d e f

data (,,,,,,) a b c d e f g

Constructors

(,,,,,,) a b c d e f g

data (,,,,,,,) a b c d e f g h

Constructors

(,,,,,,,) a b c d e f g h

data (,,,,,,,,) a b c d e f g h i

Constructors

(,,,,,,,,) a b c d e f g h i

data (,,,,,,,,,) a b c d e f g h i j

Constructors

(,,,,,,,,,) a b c d e f g h i j

data (,,,,,,,,,,) a b c d e f g h i j k

Constructors

(,,,,,,,,,,) a b c d e f g h i j k

data (,,,,,,,,,,,) a b c d e f g h i j k l

Constructors

(,,,,,,,,,,,) a b c d e f g h i j k l

data (,,,,,,,,,,,,) a b c d e f g h i j k l m

Constructors

(,,,,,,,,,,,,) a b c d e f g h i j k l m

data (,,,,,,,,,,,,,) a b c d e f g h i j k l m n

Constructors

(,,,,,,,,,,,,,) a b c d e f g h i j k l m n

data (,,,,,,,,,,,,,,) a b c d e f g h i j k l m n o

Constructors

(,,,,,,,,,,,,,,) a b c d e f g h i j k l m n o