Data.Type.Equality

Description

Definition of propositional equality (:=:). Pattern-matching on a variable of type (a :=: b) produces a proof that a ~ b.

Documentation

data a :=: b where

Propositional equality. If a :=: b is inhabited by some terminating value, then the type a is the same as the type b. To use this equality in practice, pattern-match on the a :=: b to get out the Refl constructor; in the body of the pattern-match, the compiler knows that a ~ b.

Constructors

Refl :: a :=: a

sym :: (a :=: b) -> b :=: a

Symmetry of equality

trans :: (a :=: b) -> (b :=: c) -> a :=: c

Transitivity of equality

coerce :: (a :=: b) -> a -> b

Type-safe cast, using propositional equality

liftEq :: (a :=: b) -> f a :=: f b

Lift equality into a unary type constructor

liftEq2 :: (a :=: a') -> (b :=: b') -> f a b :=: f a' b'

Lift equality into a binary type constructor

liftEq3 :: (a :=: a') -> (b :=: b') -> (c :=: c') -> f a b c :=: f a' b' c'

Lift equality into a ternary type constructor

liftEq4 :: (a :=: a') -> (b :=: b') -> (c :=: c') -> (d :=: d') -> f a b c d :=: f a' b' c' d'

Lift equality into a quaternary type constructor

lower :: (f a :=: f b) -> a :=: b

Lower equality from a parameterized type into the parameters

class EqualityT f where

This class contains types where you can learn the equality of two types from information contained in terms. Typically, only singleton types should inhabit this class.

Methods

equalsT :: f a -> f b -> Maybe (a :=: b)

Conditionally prove the equality of a and b.