Foreign.C.ConstPtr

Description

This module provides typed const pointers to foreign data. It is part of the Foreign Function Interface (FFI).

Documentation

newtype ConstPtr a

A pointer with the C const qualifier. For instance, an argument of type ConstPtr CInt would be marshalled as const int*.

While const-ness generally does not matter for ccall imports (since const and non-const pointers typically have equivalent calling conventions), it does matter for capi imports. See GHC #22043.

Since: base-4.18.0.0

Constructors

ConstPtr

Fields

unConstPtr :: Ptr a