GHC.Arr

Description

GHC's array implementation.

Documentation

class Ord a => Ix a where

The Ix class is used to map a contiguous subrange of values in a type onto integers. It is used primarily for array indexing (see the array package).

The first argument (l,u) of each of these operations is a pair specifying the lower and upper bounds of a contiguous subrange of values.

An implementation is entitled to assume the following laws about these operations:

inRange (l,u) i == elem i (range (l,u))

range (l,u) !! index (l,u) i == i, when inRange (l,u) i

map (index (l,u)) (range (l,u))) == [0..rangeSize (l,u)-1]

rangeSize (l,u) == length (range (l,u))

Minimal complete definition

range, (index | unsafeIndex), inRange

Methods

range :: (a, a) -> [a]

The list of values in the subrange defined by a bounding pair.

index :: (a, a) -> a -> Int

The position of a subscript in the subrange.

unsafeIndex :: (a, a) -> a -> Int

Like index, but without checking that the value is in range.

inRange :: (a, a) -> a -> Bool

Returns True the given subscript lies in the range defined the bounding pair.

rangeSize :: (a, a) -> Int

The size of the subrange defined by a bounding pair.

unsafeRangeSize :: (a, a) -> Int

like rangeSize, but without checking that the upper bound is in range.

data Array i e

The type of immutable non-strict (boxed) arrays with indices in i and elements in e.

Constructors

Array !i !i !Int (Array# e)

data STArray s i e

Mutable, boxed, non-strict arrays in the ST monad. The type arguments are as follows:

s: the state variable argument for the ST type

i: the index type of the array (should be an instance of Ix)

e: the element type of the array.

Constructors

STArray !i !i !Int (MutableArray# s e)

arrEleBottom :: a

array

Arguments

:: Ix i

=> (i, i)
a pair of bounds, each of the index type of the array. These bounds are the lowest and highest indices in the array, in that order. For example, a one-origin vector of length 10 has bounds (1,10), and a one-origin 10 by 10 matrix has bounds ((1,1),(10,10)).

-> [(i, e)]
a list of associations of the form (index, value). Typically, this list will be expressed as a comprehension. An association (i, x) defines the value of the array at index i to be x.

-> Array i e

Construct an array with the specified bounds and containing values for given indices within these bounds.

The array is undefined (i.e. bottom) if any index in the list is out of bounds. The Haskell 2010 Report further specifies that if any two associations in the list have the same index, the value at that index is undefined (i.e. bottom). However in GHC's implementation, the value at such an index is the value part of the last association with that index in the list.

Because the indices must be checked for these errors, array is strict in the bounds argument and in the indices of the association list, but non-strict in the values. Thus, recurrences such as the following are possible:

a = array (1,100) ((1,1) : [(i, i * a!(i-1)) | i <- [2..100]])

Not every index within the bounds of the array need appear in the association list, but the values associated with indices that do not appear will be undefined (i.e. bottom).

If, in any dimension, the lower bound is greater than the upper bound, then the array is legal, but empty. Indexing an empty array always gives an array-bounds error, but bounds still yields the bounds with which the array was constructed.

listArray :: Ix i => (i, i) -> [e] -> Array i e

Construct an array from a pair of bounds and a list of values in index order.

(!) :: Ix i => Array i e -> i -> e infixl 9

The value at the given index in an array.

safeRangeSize :: Ix i => (i, i) -> Int

negRange :: Int

safeIndex :: Ix i => (i, i) -> Int -> i -> Int

badSafeIndex

Arguments

:: Int
Index searched for.

-> Int
Max bound of the array.

-> Int
This isn't an Int. It's an exception.

Used to throw exceptions in array bounds-checking functions.

⚠ This function throws SomeException in all cases.

Examples

Expand

>>> badSafeIndex 2 5
*** Exception: Error in array index; 2 not in range [0..5)

bounds :: Array i e -> (i, i)

The bounds with which an array was constructed.

numElements :: Array i e -> Int

The number of elements in the array.

numElementsSTArray :: STArray s i e -> Int

indices :: Ix i => Array i e -> [i]

The list of indices of an array in ascending order.

elems :: Array i e -> [e]

The list of elements of an array in index order.

assocs :: Ix i => Array i e -> [(i, e)]

The list of associations of an array in index order.

accumArray

Arguments

:: Ix i

=> (e -> a -> e)
accumulating function

-> e
initial value

-> (i, i)
bounds of the array

-> [(i, a)]
association list

-> Array i e

The accumArray function deals with repeated indices in the association list using an accumulating function which combines the values of associations with the same index.

For example, given a list of values of some index type, hist produces a histogram of the number of occurrences of each index within a specified range:

hist :: (Ix a, Num b) => (a,a) -> [a] -> Array a b
hist bnds is = accumArray (+) 0 bnds [(i, 1) | i<-is, inRange bnds i]

accumArray is strict in each result of applying the accumulating function, although it is lazy in the initial value. Thus, unlike arrays built with array, accumulated arrays should not in general be recursive.

adjust :: (e -> a -> e) -> MutableArray# s e -> (Int, a) -> STRep s b -> STRep s b

(//) :: Ix i => Array i e -> [(i, e)] -> Array i e infixl 9

Constructs an array identical to the first argument except that it has been updated by the associations in the right argument. For example, if m is a 1-origin, n by n matrix, then

m//[((i,i), 0) | i <- [1..n]]

is the same matrix, except with the diagonal zeroed.

Repeated indices in the association list are handled as for array: Haskell 2010 specifies that the resulting array is undefined (i.e. bottom), but GHC's implementation uses the last association for each index.

accum :: Ix i => (e -> a -> e) -> Array i e -> [(i, a)] -> Array i e

accum f takes an array and an association list and accumulates pairs from the list into the array with the accumulating function f. Thus accumArray can be defined using accum:

accumArray f z b = accum f (array b [(i, z) | i <- range b])

accum is strict in all the results of applying the accumulation. However, it is lazy in the initial values of the array.

amap :: (a -> b) -> Array i a -> Array i b

ixmap :: (Ix i, Ix j) => (i, i) -> (i -> j) -> Array j e -> Array i e

ixmap allows for transformations on array indices. It may be thought of as providing function composition on the right with the mapping that the original array embodies.

A similar transformation of array values may be achieved using fmap from the Array instance of the Functor class.

eqArray :: (Ix i, Eq e) => Array i e -> Array i e -> Bool

cmpArray :: (Ix i, Ord e) => Array i e -> Array i e -> Ordering

cmpIntArray :: Ord e => Array Int e -> Array Int e -> Ordering

newSTArray :: Ix i => (i, i) -> e -> ST s (STArray s i e)

boundsSTArray :: STArray s i e -> (i, i)

readSTArray :: Ix i => STArray s i e -> i -> ST s e

writeSTArray :: Ix i => STArray s i e -> i -> e -> ST s ()

freezeSTArray :: STArray s i e -> ST s (Array i e)

thawSTArray :: Array i e -> ST s (STArray s i e)

foldlElems :: (b -> a -> b) -> b -> Array i a -> b

A left fold over the elements

foldlElems' :: (b -> a -> b) -> b -> Array i a -> b

A strict left fold over the elements

foldl1Elems :: (a -> a -> a) -> Array i a -> a

A left fold over the elements with no starting value

foldrElems :: (a -> b -> b) -> b -> Array i a -> b

A right fold over the elements

foldrElems' :: (a -> b -> b) -> b -> Array i a -> b

A strict right fold over the elements

foldr1Elems :: (a -> a -> a) -> Array i a -> a

A right fold over the elements with no starting value

Unsafe operations

fill :: MutableArray# s e -> (Int, e) -> STRep s a -> STRep s a

done :: i -> i -> Int -> MutableArray# s e -> STRep s (Array i e)

unsafeArray :: Ix i => (i, i) -> [(Int, e)] -> Array i e

unsafeArray' :: (i, i) -> Int -> [(Int, e)] -> Array i e

lessSafeIndex :: Ix i => (i, i) -> Int -> i -> Int

unsafeAt :: Array i e -> Int -> e

unsafeReplace :: Array i e -> [(Int, e)] -> Array i e

unsafeAccumArray :: Ix i => (e -> a -> e) -> e -> (i, i) -> [(Int, a)] -> Array i e

unsafeAccumArray' :: (e -> a -> e) -> e -> (i, i) -> Int -> [(Int, a)] -> Array i e

unsafeAccum :: (e -> a -> e) -> Array i e -> [(Int, a)] -> Array i e

unsafeReadSTArray :: STArray s i e -> Int -> ST s e

unsafeWriteSTArray :: STArray s i e -> Int -> e -> ST s ()

unsafeFreezeSTArray :: STArray s i e -> ST s (Array i e)

unsafeThawSTArray :: Array i e -> ST s (STArray s i e)