GHC.Clock

Documentation

getMonotonicTime :: IO Double

Return monotonic time in seconds, since some unspecified starting point

Since: base-4.11.0.0

getMonotonicTimeNSec :: IO Word64

Return monotonic time in nanoseconds, since some unspecified starting point

Since: base-4.11.0.0