GHC.Conc.Windows

Description

Windows I/O manager

Documentation

ensureIOManagerIsRunning :: IO ()

Waiting

threadDelay :: Int -> IO ()

Suspends the current thread for a given number of microseconds (GHC only).

There is no guarantee that the thread will be rescheduled promptly when the delay has expired, but the thread will never continue to run earlier than specified.

registerDelay :: Int -> IO (TVar Bool)

Set the value of returned TVar to True after a given number of microseconds. The caveats associated with threadDelay also apply.

Miscellaneous

asyncRead :: Int -> Int -> Int -> Ptr a -> IO (Int, Int)

asyncWrite :: Int -> Int -> Int -> Ptr a -> IO (Int, Int)

asyncDoProc :: FunPtr (Ptr a -> IO Int) -> Ptr a -> IO Int

asyncReadBA :: Int -> Int -> Int -> Int -> MutableByteArray# RealWorld -> IO (Int, Int)

asyncWriteBA :: Int -> Int -> Int -> Int -> MutableByteArray# RealWorld -> IO (Int, Int)

data ConsoleEvent

Constructors

ControlC

Break

Close

Logoff

Shutdown

win32ConsoleHandler :: MVar (ConsoleEvent -> IO ())

toWin32ConsoleEvent :: (Eq a, Num a) => a -> Maybe ConsoleEvent