GHC.Desugar

Description

Deprecated: GHC.Desugar is deprecated and will be removed in GHC 9.14.(>>>) should be imported from Control.Arrow.AnnotationWrapper is internal to GHC and should not be used externally.

Support code for desugaring in GHC

The API of this module is unstable and not meant to be consumed by the general public. If you absolutely must depend on it, make sure to use a tight upper bound, e.g., base < 4.X rather than base < 5, because the interface can change rapidly without much warning.

Documentation

(>>>) :: Arrow arr => forall a b c. arr a b -> arr b c -> arr a c

data AnnotationWrapper

Constructors

Data a => AnnotationWrapper a

toAnnotationWrapper :: Data a => a -> AnnotationWrapper