GHC.Encoding.UTF8

Description

Simple UTF-8 codecs supporting non-streaming encoding/decoding. For encoding where codepoints may be broken across buffers, see GHC.IO.Encoding.UTF8.

This is one of several UTF-8 implementations provided by GHC; see Note [GHC's many UTF-8 implementations] in GHC.Encoding.UTF8 for an overview.

Decoding single characters

utf8DecodeCharAddr# :: Addr# -> Int# -> (# Char#, Int# #)

Decode a single character at the given Addr#.

utf8DecodeCharPtr :: Ptr Word8 -> (Char, Int)

Decode a single codepoint starting at the given Ptr.

utf8DecodeCharByteArray# :: ByteArray# -> Int# -> (# Char#, Int# #)

Decode a single codepoint starting at the given byte offset into a ByteArray#.

Decoding strings

utf8DecodeByteArray# :: ByteArray# -> [Char]

utf8DecodeForeignPtr :: ForeignPtr Word8 -> Int -> Int -> [Char]

Counting characters

utf8CountCharsByteArray# :: ByteArray# -> Int

Comparison

utf8CompareByteArray# :: ByteArray# -> ByteArray# -> Ordering

Encoding strings

utf8EncodePtr :: Ptr Word8 -> String -> IO ()

utf8EncodeByteArray# :: String -> ByteArray#

utf8EncodedLength :: String -> Int