GHC.Environment

Documentation

getFullArgs :: IO [String]

Computation getFullArgs is the "raw" version of getArgs, similar to argv in other languages. It returns a list of the program's command line arguments, starting with the program name, and including those normally eaten by the RTS (+RTS ... -RTS).