GHC.Err

Description

The GHC.Err module defines the code for the wired-in error functions, which have a special type in the compiler (with "open tyvars").

We cannot define these functions in a module where they might be used (e.g., GHC.Base), because the magical wired-in type will get confused with what the typechecker figures out.

Documentation

absentErr :: a

Used for compiler-generated error message; encoding saves bytes of string junk.

error :: forall (r :: RuntimeRep). forall (a :: TYPE r). HasCallStack => [Char] -> a

error stops execution and displays an error message.

errorWithoutStackTrace :: forall (r :: RuntimeRep). forall (a :: TYPE r). [Char] -> a

A variant of error that does not produce a stack trace.

Since: base-4.9.0.0

undefined :: forall (r :: RuntimeRep). forall (a :: TYPE r). HasCallStack => a

A special case of error. It is expected that compilers will recognize this and insert error messages which are more appropriate to the context in which undefined appears.