GHC.Event

Description

This module provides scalable event notification for file descriptors and timeouts.

This module should be considered GHC internal.

---------------------------------------------------------------------------

Types

data EventManager

The event manager state.

data TimerManager

The event manager state.

Creation

getSystemEventManager :: IO (Maybe EventManager)

Retrieve the system event manager for the capability on which the calling thread is running.

This function always returns Just the current thread's event manager when using the threaded RTS and Nothing otherwise.

new :: IO EventManager

Create a new event manager.

getSystemTimerManager :: IO TimerManager

Registering interest in I/O events

data Event

An I/O event.

evtRead :: Event

Data is available to be read.

evtWrite :: Event

The file descriptor is ready to accept a write.

type IOCallback = FdKey -> Event -> IO ()

Callback invoked on I/O events.

data FdKey

A file descriptor registration cookie.

data Lifetime

The lifetime of an event registration.

Since: base-4.8.1.0

Constructors

OneShot
the registration will be active for only one event

MultiShot
the registration will trigger multiple times

registerFd :: EventManager -> IOCallback -> Fd -> Event -> Lifetime -> IO FdKey

registerFd mgr cb fd evs lt registers interest in the events evs on the file descriptor fd for lifetime lt. cb is called for each event that occurs. Returns a cookie that can be handed to unregisterFd.

unregisterFd :: EventManager -> FdKey -> IO ()

Drop a previous file descriptor registration.

unregisterFd_ :: EventManager -> FdKey -> IO Bool

Drop a previous file descriptor registration, without waking the event manager thread. The return value indicates whether the event manager ought to be woken.

closeFd :: EventManager -> (Fd -> IO ()) -> Fd -> IO ()

Close a file descriptor in a race-safe way. It might block, although for a very short time; and thus it is interruptible by asynchronous exceptions.

Registering interest in timeout events

type TimeoutCallback = IO ()

Warning: since the TimeoutCallback is called from the I/O manager, it must not throw an exception or block for a long period of time. In particular, be wary of throwTo and killThread: if the target thread is making a foreign call, these functions will block until the call completes.

data TimeoutKey

A timeout registration cookie.

registerTimeout :: TimerManager -> Int -> TimeoutCallback -> IO TimeoutKey

Register a timeout in the given number of microseconds. The returned TimeoutKey can be used to later unregister or update the timeout. The timeout is automatically unregistered after the given time has passed.

Be careful not to exceed maxBound :: Int, which on 32-bit machines is only 2147483647 μs, less than 36 minutes.

updateTimeout :: TimerManager -> TimeoutKey -> Int -> IO ()

Update an active timeout to fire in the given number of microseconds.

Be careful not to exceed maxBound :: Int, which on 32-bit machines is only 2147483647 μs, less than 36 minutes.

unregisterTimeout :: TimerManager -> TimeoutKey -> IO ()

Unregister an active timeout.