GHC.Event.Internal

Event back end

data Backend

Event notification backend.

backend :: (a -> Maybe Timeout -> (Fd -> Event -> IO ()) -> IO Int) -> (a -> Fd -> Event -> Event -> IO Bool) -> (a -> Fd -> Event -> IO Bool) -> (a -> IO ()) -> a -> Backend

delete :: Backend -> IO ()

poll :: Backend -> Maybe Timeout -> (Fd -> Event -> IO ()) -> IO Int

modifyFd :: Backend -> Fd -> Event -> Event -> IO Bool

Returns True if the modification succeeded. Returns False if this backend does not support event notifications on this type of file.

modifyFdOnce :: Backend -> Fd -> Event -> IO Bool

Returns True if the modification succeeded. Returns False if this backend does not support event notifications on this type of file.

Event type

data Event

An I/O event.

evtRead :: Event

Data is available to be read.

evtWrite :: Event

The file descriptor is ready to accept a write.

evtClose :: Event

Another thread closed the file descriptor.

evtNothing :: Event

eventIs :: Event -> Event -> Bool

Lifetimes

data Lifetime

The lifetime of an event registration.

Since: base-4.8.1.0

Constructors

OneShot
the registration will be active for only one event

MultiShot
the registration will trigger multiple times

data EventLifetime

A pair of an event and lifetime

Here we encode the event in the bottom three bits and the lifetime in the fourth bit.

eventLifetime :: Event -> Lifetime -> EventLifetime

elLifetime :: EventLifetime -> Lifetime

elEvent :: EventLifetime -> Event

Timeout type

data Timeout

A type alias for timeouts, specified in nanoseconds.

Constructors

Timeout !Word64

Forever

Helpers

throwErrnoIfMinus1NoRetry :: (Eq a, Num a) => String -> IO a -> IO a

Throw an IOError corresponding to the current value of getErrno if the result value of the IO action is -1 and getErrno is not eINTR. If the result value is -1 and getErrno returns eINTR 0 is returned. Otherwise the result value is returned.

exchangePtr :: Ptr (Ptr a) -> Ptr a -> IO (Ptr a)

exchangePtr pptr x swaps the pointer pointed to by pptr with the value x, returning the old value.