GHC.Event.TimeOut

Description

Common Timer definitions shared between WinIO and RIO.

Documentation

type TimeoutQueue = PSQ TimeoutCallback

A priority search queue, with timeouts as priorities.

type TimeoutCallback = IO ()

Warning: since the TimeoutCallback is called from the I/O manager, it must not throw an exception or block for a long period of time. In particular, be wary of throwTo and killThread: if the target thread is making a foreign call, these functions will block until the call completes.

type TimeoutEdit = TimeoutQueue -> TimeoutQueue

An edit to apply to a TimeoutQueue.

newtype TimeoutKey

A timeout registration cookie.

Constructors

TK Unique