GHC.Event.Windows.Clock

Documentation

data Clock

Monotonic clock

type Seconds = Double

getTime :: Clock -> IO Seconds

Get the current time, in seconds since some fixed time in the past.

getClock :: IO Clock

Figure out what time API to use, and return a Clock for accessing it.

Specific implementations

queryPerformanceCounter :: IO (Maybe Clock)

getTickCount64 :: IO Clock