GHC.Float.RealFracMethods

Description

Methods for the RealFrac instances for Float and Double, with specialised versions for Int.

Moved to their own module to not bloat GHC.Float further.

The API of this module is unstable and not meant to be consumed by the general public. If you absolutely must depend on it, make sure to use a tight upper bound, e.g., base < 4.X rather than base < 5, because the interface can change rapidly without much warning.

Double methods

Integer results

properFractionDoubleInteger :: Double -> (Integer, Double)

truncateDoubleInteger :: Double -> Integer

floorDoubleInteger :: Double -> Integer

ceilingDoubleInteger :: Double -> Integer

roundDoubleInteger :: Double -> Integer

Int results

properFractionDoubleInt :: Double -> (Int, Double)

floorDoubleInt :: Double -> Int

ceilingDoubleInt :: Double -> Int

roundDoubleInt :: Double -> Int

Double/Int conversions, wrapped primops

double2Int :: Double -> Int

int2Double :: Int -> Double

Float methods

Integer results

properFractionFloatInteger :: Float -> (Integer, Float)

truncateFloatInteger :: Float -> Integer

floorFloatInteger :: Float -> Integer

ceilingFloatInteger :: Float -> Integer

roundFloatInteger :: Float -> Integer

Int results

properFractionFloatInt :: Float -> (Int, Float)

floorFloatInt :: Float -> Int

ceilingFloatInt :: Float -> Int

roundFloatInt :: Float -> Int

Float/Int conversions, wrapped primops

float2Int :: Float -> Int

int2Float :: Int -> Float