GHC.Foreign

Description

Foreign marshalling support for CStrings with configurable encodings

C strings with a configurable encoding

type CString = Ptr CChar

A C string is a reference to an array of C characters terminated by NUL.

type CStringLen = (Ptr CChar, Int)

A string with explicit length information in bytes instead of a terminating NUL (allowing NUL characters in the middle of the string).

peekCString :: TextEncoding -> CString -> IO String

Marshal a NUL terminated C string into a Haskell string.

peekCStringLen :: TextEncoding -> CStringLen -> IO String

Marshal a C string with explicit length into a Haskell string.

newCString :: TextEncoding -> String -> IO CString

Marshal a Haskell string into a NUL terminated C string.

the Haskell string may not contain any NUL characters

new storage is allocated for the C string and must be explicitly freed using free or finalizerFree.

newCStringLen :: TextEncoding -> String -> IO CStringLen

Marshal a Haskell string into a C string (ie, character array) with explicit length information.

new storage is allocated for the C string and must be explicitly freed using free or finalizerFree.

withCString :: TextEncoding -> String -> (CString -> IO a) -> IO a

Marshal a Haskell string into a NUL terminated C string using temporary storage.

the Haskell string may not contain any NUL characters

the memory is freed when the subcomputation terminates (either normally or via an exception), so the pointer to the temporary storage must not be used after this.

withCStringLen :: TextEncoding -> String -> (CStringLen -> IO a) -> IO a

Marshal a Haskell string into a C string (ie, character array) in temporary storage, with explicit length information.

the memory is freed when the subcomputation terminates (either normally or via an exception), so the pointer to the temporary storage must not be used after this.

withCStringsLen :: TextEncoding -> [String] -> (Int -> Ptr CString -> IO a) -> IO a

Marshal a list of Haskell strings into an array of NUL terminated C strings using temporary storage.

the Haskell strings may not contain any NUL characters

the memory is freed when the subcomputation terminates (either normally or via an exception), so the pointer to the temporary storage must not be used after this.

charIsRepresentable :: TextEncoding -> Char -> IO Bool

Determines whether a character can be accurately encoded in a CString.

Pretty much anyone who uses this function is in a state of sin because whether or not a character is encodable will, in general, depend on the context in which it occurs.