GHC.IO.Handle.Windows

Description

Handle operations implemented by Windows native handles

Documentation

stdin :: Handle

A handle managing input from the Haskell program's standard input channel.

stdout :: Handle

A handle managing output to the Haskell program's standard output channel.

stderr :: Handle

A handle managing output to the Haskell program's standard error channel.

openFile :: FilePath -> IOMode -> IO Handle

Computation openFile file mode allocates and returns a new, open handle to manage the file file. It manages input if mode is ReadMode, output if mode is WriteMode or AppendMode, and both input and output if mode is ReadWriteMode.

If the file does not exist and it is opened for output, it should be created as a new file. If mode is WriteMode and the file already exists, then it should be truncated to zero length. Some operating systems delete empty files, so there is no guarantee that the file will exist following an openFile with mode WriteMode unless it is subsequently written to successfully. The handle is positioned at the end of the file if mode is AppendMode, and otherwise at the beginning (in which case its internal position is 0). The initial buffer mode is implementation-dependent.

This operation may fail with:

isAlreadyInUseError if the file is already open and cannot be reopened;

isDoesNotExistError if the file does not exist; or

isPermissionError if the user does not have permission to open the file.

Note: if you will be working with files containing binary data, you'll want to be using openBinaryFile.

openBinaryFile :: FilePath -> IOMode -> IO Handle

Like openFile, but open the file in binary mode. On Windows, reading a file in text mode (which is the default) will translate CRLF to LF, and writing will translate LF to CRLF. This is usually what you want with text files. With binary files this is undesirable; also, as usual under Microsoft operating systems, text mode treats control-Z as EOF. Binary mode turns off all special treatment of end-of-line and end-of-file characters. (See also hSetBinaryMode.)

openFileBlocking :: FilePath -> IOMode -> IO Handle

Like openFile, but opens the file in ordinary blocking mode. This can be useful for opening a FIFO for writing: if we open in non-blocking mode then the open will fail if there are no readers, whereas a blocking open will block until a reader appear.

Since: base-4.4.0.0

handleToHANDLE :: Handle -> IO HANDLE

Turn an existing Handle into a Win32 HANDLE. This function throws an IOError if the Handle does not reference a HANDLE

mkHandleFromHANDLE :: (RawIO dev, IODevice dev, BufferedIO dev, Typeable dev) => dev -> IODeviceType -> FilePath -> IOMode -> Maybe TextEncoding -> IO Handle