GHC.IO.SubSystem

Description

The IoSubSystem control interface. These methods can be used to disambiguate between the two operations.

The API of this module is unstable and not meant to be consumed by the general public. If you absolutely must depend on it, make sure to use a tight upper bound, e.g., base < 4.X rather than base < 5, because the interface can change rapidly without much warning.

Documentation

withIoSubSystem :: (IoSubSystem -> IO a) -> IO a

withIoSubSystem' :: (IoSubSystem -> a) -> a

whenIoSubSystem :: IoSubSystem -> IO () -> IO ()

ioSubSystem :: IoSubSystem

data IoSubSystem

The I/O SubSystem to use in the program.

@since base-4.9.0.0

Constructors

IoPOSIX
Use a POSIX I/O Sub-System

IoNative
Use platform native Sub-System. For unix OSes this is the same as IoPOSIX, but on Windows this means use the Windows native APIs for I/O, including IOCP and RIO.

conditional :: a -> a -> a

Conditionally execute an action depending on the configured I/O subsystem. On POSIX systems always execute the first action. On Windows execute the second action if WINIO as active, otherwise fall back to the first action.

(<!>) :: a -> a -> a infixl 7

Infix version of conditional. posix ! windows == conditional posix windows

isWindowsNativeIO :: Bool