GHC.Int

Description

The sized integral datatypes, Int8, Int16, Int32, and Int64.

Documentation

data Int

A fixed-precision integer type with at least the range [-2^29 .. 2^29-1]. The exact range for a given implementation can be determined by using minBound and maxBound from the Bounded class.

Constructors

I# Int#

data Int8

8-bit signed integer type

Constructors

I8# Int8#

data Int16

16-bit signed integer type

Constructors

I16# Int16#

data Int32

32-bit signed integer type

Constructors

I32# Int32#

data Int64

64-bit signed integer type

Constructors

I64# Int64#

uncheckedIShiftL64# :: Int64# -> Int# -> Int64#

uncheckedIShiftRA64# :: Int64# -> Int# -> Int64#

shiftRLInt8# :: Int8# -> Int# -> Int8#

shiftRLInt16# :: Int16# -> Int# -> Int16#

shiftRLInt32# :: Int32# -> Int# -> Int32#

Equality operators

See GHC.Classes#matching_overloaded_methods_in_rules

eqInt :: Int -> Int -> Bool

neInt :: Int -> Int -> Bool

gtInt :: Int -> Int -> Bool

geInt :: Int -> Int -> Bool

ltInt :: Int -> Int -> Bool

leInt :: Int -> Int -> Bool

eqInt8 :: Int8 -> Int8 -> Bool

neInt8 :: Int8 -> Int8 -> Bool

gtInt8 :: Int8 -> Int8 -> Bool

geInt8 :: Int8 -> Int8 -> Bool

ltInt8 :: Int8 -> Int8 -> Bool

leInt8 :: Int8 -> Int8 -> Bool

eqInt16 :: Int16 -> Int16 -> Bool

neInt16 :: Int16 -> Int16 -> Bool

gtInt16 :: Int16 -> Int16 -> Bool

geInt16 :: Int16 -> Int16 -> Bool

ltInt16 :: Int16 -> Int16 -> Bool

leInt16 :: Int16 -> Int16 -> Bool

eqInt32 :: Int32 -> Int32 -> Bool

neInt32 :: Int32 -> Int32 -> Bool

gtInt32 :: Int32 -> Int32 -> Bool

geInt32 :: Int32 -> Int32 -> Bool

ltInt32 :: Int32 -> Int32 -> Bool

leInt32 :: Int32 -> Int32 -> Bool

eqInt64 :: Int64 -> Int64 -> Bool

neInt64 :: Int64 -> Int64 -> Bool

gtInt64 :: Int64 -> Int64 -> Bool

geInt64 :: Int64 -> Int64 -> Bool

ltInt64 :: Int64 -> Int64 -> Bool

leInt64 :: Int64 -> Int64 -> Bool