GHC.Natural

Description

Compatibility module for pre ghc-bignum code.

Documentation

data Natural where

Natural number

Invariant: numbers <= 0xffffffffffffffff use the NS constructor

Bundled Patterns

pattern NatS# :: Word# -> Natural

pattern NatJ# :: BigNat -> Natural

data BigNat

A lifted BigNat

Represented as an array of limbs (Word#) stored in little-endian order (Word# themselves use machine order).

Invariant (canonical representation): higher Word# is non-zero.

As a consequence, zero is represented with a WordArray# whose size is 0.

Constructors

BN#

Fields

unBigNat :: BigNat#

mkNatural :: [Word] -> Natural

Construct Natural value from list of Words.

isValidNatural :: Natural -> Bool

Test whether all internal invariants are satisfied by Natural value

This operation is mostly useful for test-suites and/or code which constructs Integer values directly.

Since: base-4.8.0.0

Arithmetic

plusNatural :: Natural -> Natural -> Natural

Natural Addition

minusNatural :: Natural -> Natural -> Natural

Natural subtraction. May throw Underflow.

minusNaturalMaybe :: Natural -> Natural -> Maybe Natural

Natural subtraction. Returns Nothings for non-positive results.

Since: base-4.8.0.0

timesNatural :: Natural -> Natural -> Natural

Natural multiplication

negateNatural :: Natural -> Natural

signumNatural :: Natural -> Natural

quotRemNatural :: Natural -> Natural -> (Natural, Natural)

quotNatural :: Natural -> Natural -> Natural

remNatural :: Natural -> Natural -> Natural

gcdNatural :: Natural -> Natural -> Natural

Compute greatest common divisor.

lcmNatural :: Natural -> Natural -> Natural

Compute least common multiple.

Bits

andNatural :: Natural -> Natural -> Natural

orNatural :: Natural -> Natural -> Natural

xorNatural :: Natural -> Natural -> Natural

bitNatural :: Int# -> Natural

testBitNatural :: Natural -> Int -> Bool

popCountNatural :: Natural -> Int

shiftLNatural :: Natural -> Int -> Natural

shiftRNatural :: Natural -> Int -> Natural

Conversions

naturalToInteger :: Natural -> Integer

Since: base-4.12.0.0

naturalToWord :: Natural -> Word

naturalToWordMaybe :: Natural -> Maybe Word

Try downcasting Natural to Word value. Returns Nothing if value doesn't fit in Word.

Since: base-4.8.0.0

wordToNatural :: Word -> Natural

Construct Natural from Word value.

Since: base-4.8.0.0

wordToNatural# :: Word -> Natural

naturalFromInteger :: Integer -> Natural

Since: base-4.10.0.0

Modular arithmetic

powModNatural :: Natural -> Natural -> Natural -> Natural

"powModNatural b e m" computes base b raised to exponent e modulo m.

Since: base-4.8.0.0