GHC.Pack

Description

⚠ Warning: Starting base-4.18, this module is being deprecated. See https://gitlab.haskell.org/ghc/ghc/-/issues/21461 for more information.

This module provides a small set of low-level functions for packing and unpacking a chunk of bytes. Used by code emitted by the compiler plus the prelude libraries.

The programmer level view of packed strings is provided by a GHC system library PackedString.

Documentation

packCString# :: [Char] -> ByteArray#

unpackCString :: Ptr a -> [Char]

unpackCString# :: Addr# -> [Char]

unpackNBytes# :: Addr# -> Int# -> [Char]

unpackFoldrCString# :: Addr# -> (Char -> a -> a) -> a -> a

unpackAppendCString# :: Addr# -> [Char] -> [Char]