GHC.Profiling

Description

Since: base-4.7.0.0

Cost Centre Profiling

startProfTimer :: IO ()

Start attributing ticks to cost centres. This is called by the RTS on startup.

Since: base-4.7.0.0

stopProfTimer :: IO ()

Stop attributing ticks to cost centres. Allocations will still be attributed.

Since: base-4.7.0.0

Heap Profiling

startHeapProfTimer :: IO ()

Start heap profiling. This is called normally by the RTS on start-up, but can be disabled using the rts flag `--no-automatic-heap-samples` Note: This won't do anything unless you also specify a profiling mode on the command line using the normal RTS options.

Since: base-4.16.0.0

stopHeapProfTimer :: IO ()

Stop heap profiling. Note: This won't do anything unless you also specify a profiling mode on the command line using the normal RTS options.

Since: base-4.16.0.0

requestHeapCensus :: IO ()

Request a heap census on the next context switch. The census can be requested whether or not the heap profiling timer is running. Note: This won't do anything unless you also specify a profiling mode on the command line using the normal RTS options.

Since: base-4.16.0.0