GHC.RTS.Flags

Description

Accessors to GHC RTS flags. Descriptions of flags can be seen in GHC User's Guide, or by running RTS help message using +RTS --help.

Since: base-4.8.0.0

Documentation

type RtsTime = Word64

RtsTime is defined as a StgWord64 in stg/Types.h

Since: base-4.8.2.0

data RTSFlags

Parameters of the runtime system

Since: base-4.8.0.0

Constructors

RTSFlags

Fields

gcFlags :: GCFlags

concurrentFlags :: ConcFlags

miscFlags :: MiscFlags

debugFlags :: DebugFlags

costCentreFlags :: CCFlags

profilingFlags :: ProfFlags

traceFlags :: TraceFlags

tickyFlags :: TickyFlags

parFlags :: ParFlags

data GiveGCStats

Should we produce a summary of the garbage collector statistics after the program has exited?

Since: base-4.8.2.0

Constructors

NoGCStats

CollectGCStats

OneLineGCStats

SummaryGCStats

VerboseGCStats

data GCFlags

Parameters of the garbage collector.

Since: base-4.8.0.0

Constructors

GCFlags

Fields

statsFile :: Maybe FilePath

giveStats :: GiveGCStats

maxStkSize :: Word32

initialStkSize :: Word32

stkChunkSize :: Word32

stkChunkBufferSize :: Word32

maxHeapSize :: Word32

minAllocAreaSize :: Word32

largeAllocLim :: Word32

nurseryChunkSize :: Word32

minOldGenSize :: Word32

heapSizeSuggestion :: Word32

heapSizeSuggestionAuto :: Bool

oldGenFactor :: Double

returnDecayFactor :: Double

pcFreeHeap :: Double

generations :: Word32

squeezeUpdFrames :: Bool

compact :: Bool

True = "compact all the time"

compactThreshold :: Double

sweep :: Bool

use "mostly mark-sweep" instead of copying for the oldest generation

ringBell :: Bool

idleGCDelayTime :: RtsTime

doIdleGC :: Bool

heapBase :: Word

address to ask the OS for memory

allocLimitGrace :: Word

numa :: Bool

numaMask :: Word

data ConcFlags

Parameters concerning context switching

Since: base-4.8.0.0

Constructors

ConcFlags

Fields

ctxtSwitchTime :: RtsTime

ctxtSwitchTicks :: Int

data MiscFlags

Miscellaneous parameters

Since: base-4.8.0.0

Constructors

MiscFlags

Fields

tickInterval :: RtsTime

installSignalHandlers :: Bool

installSEHHandlers :: Bool

generateCrashDumpFile :: Bool

generateStackTrace :: Bool

machineReadable :: Bool

disableDelayedOsMemoryReturn :: Bool

internalCounters :: Bool

linkerAlwaysPic :: Bool

linkerMemBase :: Word

address to ask the OS for memory for the linker, 0 ==> off

ioManager :: IoSubSystem

numIoWorkerThreads :: Word32

data DebugFlags

Flags to control debugging output & extra checking in various subsystems.

Since: base-4.8.0.0

Constructors

DebugFlags

Fields

scheduler :: Bool

s

interpreter :: Bool

i

weak :: Bool

w

gccafs :: Bool

G

gc :: Bool

g

nonmoving_gc :: Bool

n

block_alloc :: Bool

b

sanity :: Bool

S

stable :: Bool

t

prof :: Bool

p

linker :: Bool

l the object linker

apply :: Bool

a

stm :: Bool

m

squeeze :: Bool

z stack squeezing & lazy blackholing

hpc :: Bool

c coverage

sparks :: Bool

r

data DoCostCentres

Should the RTS produce a cost-center summary?

Since: base-4.8.2.0

Constructors

CostCentresNone

CostCentresSummary

CostCentresVerbose

CostCentresAll

CostCentresJSON

data CCFlags

Parameters pertaining to the cost-center profiler.

Since: base-4.8.0.0

Constructors

CCFlags

Fields

doCostCentres :: DoCostCentres

profilerTicks :: Int

msecsPerTick :: Int

data DoHeapProfile

What sort of heap profile are we collecting?

Since: base-4.8.2.0

Constructors

NoHeapProfiling

HeapByCCS

HeapByMod

HeapByDescr

HeapByType

HeapByRetainer

HeapByLDV

HeapByClosureType

HeapByInfoTable

data ProfFlags

Parameters of the cost-center profiler

Since: base-4.8.0.0

Constructors

ProfFlags

Fields

doHeapProfile :: DoHeapProfile

heapProfileInterval :: RtsTime

time between samples

heapProfileIntervalTicks :: Word

ticks between samples (derived)

startHeapProfileAtStartup :: Bool

showCCSOnException :: Bool

maxRetainerSetSize :: Word

ccsLength :: Word

modSelector :: Maybe String

descrSelector :: Maybe String

typeSelector :: Maybe String

ccSelector :: Maybe String

ccsSelector :: Maybe String

retainerSelector :: Maybe String

bioSelector :: Maybe String

data DoTrace

Is event tracing enabled?

Since: base-4.8.2.0

Constructors

TraceNone
no tracing

TraceEventLog
send tracing events to the event log

TraceStderr
send tracing events to stderr

data TraceFlags

Parameters pertaining to event tracing

Since: base-4.8.0.0

Constructors

TraceFlags

Fields

tracing :: DoTrace

timestamp :: Bool

show timestamp in stderr output

traceScheduler :: Bool

trace scheduler events

traceGc :: Bool

trace GC events

traceNonmovingGc :: Bool

trace nonmoving GC heap census samples

sparksSampled :: Bool

trace spark events by a sampled method

sparksFull :: Bool

trace spark events 100% accurately

user :: Bool

trace user events (emitted from Haskell code)

data TickyFlags

Parameters pertaining to ticky-ticky profiler

Since: base-4.8.0.0

Constructors

TickyFlags

Fields

showTickyStats :: Bool

tickyFile :: Maybe FilePath

data ParFlags

Parameters pertaining to parallelism

Since: base-4.8.0.0

Constructors

ParFlags

Fields

nCapabilities :: Word32

migrate :: Bool

maxLocalSparks :: Word32

parGcEnabled :: Bool

parGcGen :: Word32

parGcLoadBalancingEnabled :: Bool

parGcLoadBalancingGen :: Word32

parGcNoSyncWithIdle :: Word32

parGcThreads :: Word32

setAffinity :: Bool

data IoSubSystem

The I/O SubSystem to use in the program.

Since: base-4.9.0.0

Constructors

IoPOSIX
Use a POSIX I/O Sub-System

IoNative
Use platform native Sub-System. For unix OSes this is the same as IoPOSIX, but on Windows this means use the Windows native APIs for I/O, including IOCP and RIO.

getRTSFlags :: IO RTSFlags

getGCFlags :: IO GCFlags

getConcFlags :: IO ConcFlags

getMiscFlags :: IO MiscFlags

getIoManagerFlag :: IO IoSubSystem

Needed to optimize support for different IO Managers on Windows. See Note [The need for getIoManagerFlag]

getDebugFlags :: IO DebugFlags

getCCFlags :: IO CCFlags

getProfFlags :: IO ProfFlags

getTraceFlags :: IO TraceFlags

getTickyFlags :: IO TickyFlags

getParFlags :: IO ParFlags