GHC.Records

Description

This module defines the HasField class used by the OverloadedRecordFields extension. See the <https://gitlab.haskell.org/ghc/ghc/wikis/records/overloaded-record-fields wiki page> for more details.

Documentation

class HasField x r a | x r -> a where

Constraint representing the fact that the field x belongs to the record type r and has field type a. This will be solved automatically, but manual instances may be provided as well.

Methods

getField :: r -> a

Selector function to extract the field from the record.