GHC.STRef

Description

References in the ST monad.

Documentation

data STRef s a

a value of type STRef s a is a mutable variable in state thread s, containing a value of type a

>>> :{
runST (do
    ref <- newSTRef "hello"
    x <- readSTRef ref
    writeSTRef ref (x ++ "world")
    readSTRef ref )
:}
"helloworld"

Constructors

STRef (MutVar# s a)

newSTRef :: a -> ST s (STRef s a)

Build a new STRef in the current state thread

readSTRef :: STRef s a -> ST s a

Read the value of an STRef

writeSTRef :: STRef s a -> a -> ST s ()

Write a new value into an STRef