GHC.Stack.CCS

Description

Access to GHC's call-stack simulation

Since: base-4.5.0.0

Call stacks

currentCallStack :: IO [String]

Returns a [String] representing the current call stack. This can be useful for debugging.

The implementation uses the call-stack simulation maintained by the profiler, so it only works if the program was compiled with -prof and contains suitable SCC annotations (e.g. by using -fprof-auto). Otherwise, the list returned is likely to be empty or uninformative.

Since: base-4.5.0.0

whoCreated :: a -> IO [String]

Get the stack trace attached to an object.

Since: base-4.5.0.0

Internals

data CostCentreStack

A cost-centre stack from GHC's cost-center profiler.

data CostCentre

A cost-centre from GHC's cost-center profiler.

getCurrentCCS :: dummy -> IO (Ptr CostCentreStack)

Returns the current CostCentreStack (value is nullPtr if the current program was not compiled with profiling support). Takes a dummy argument which can be used to avoid the call to getCurrentCCS being floated out by the simplifier, which would result in an uninformative stack (CAF).

getCCSOf :: a -> IO (Ptr CostCentreStack)

Get the CostCentreStack associated with the given value.

clearCCS :: IO a -> IO a

Run a computation with an empty cost-center stack. For example, this is used by the interpreter to run an interpreted computation without the call stack showing that it was invoked from GHC.

ccsCC :: Ptr CostCentreStack -> IO (Ptr CostCentre)

Get the CostCentre at the head of a CostCentreStack.

ccsParent :: Ptr CostCentreStack -> IO (Ptr CostCentreStack)

Get the tail of a CostCentreStack.

ccLabel :: Ptr CostCentre -> IO CString

Get the label of a CostCentre.

ccModule :: Ptr CostCentre -> IO CString

Get the module of a CostCentre.

ccSrcSpan :: Ptr CostCentre -> IO CString

Get the source span of a CostCentre.

ccsToStrings :: Ptr CostCentreStack -> IO [String]

Format a CostCentreStack as a list of lines.

renderStack :: [String] -> String