GHC.TypeLits

Description

This module is an internal GHC module. It declares the constants used in the implementation of type-level natural numbers. The programmer interface for working with type-level naturals should be defined in a separate library.

Kinds

data Nat

(Kind) This is the kind of type-level natural numbers.

data Symbol

(Kind) This is the kind of type-level symbols.

Linking type and value level

data family Sing n

class SingI a where

The class SingI provides a "smart" constructor for singleton types. There are built-in instances for the singleton types corresponding to type literals.

Methods

sing :: Sing a

The only value of type Sing a

class kparam ~ KindParam => SingE kparam where

A class that converts singletons of a given kind into values of some representation type (i.e., we forget the extra information carried by the singletons, and convert them to ordinary values).

Note that fromSing is overloaded based on the kind of the values and not their type---all types of a given kind are processed by the same instances.

Associated Types

type DemoteRep kparam :: *

Methods

fromSing :: Sing (a :: k) -> DemoteRep kparam

class (SingI a, SingE (KindOf a)) => SingRep a

A convenience class, useful when we need to both introduce and eliminate a given singleton value. Users should never need to define instances of this classes.

singByProxy :: SingI n => proxy n -> Sing n

A convenience function to create a singleton value, when we have a proxy argument in scope.

data SomeSing where

Constructors

SomeSing :: SingI (n :: k) => proxy n -> SomeSing (kp :: KindIs k)

class kp ~ KindParam => ToSing kp where

The class of function that can promote a representation value into a singleton. Like SingE, this class overloads based on a *kind*. The method returns Maybe because sometimes the representation type contains more values than are supported by the singletons.

Methods

toSing :: DemoteRep kp -> Maybe (SomeSing kp)

type SomeNat = SomeSing (KindParam :: KindIs Nat)

A definition of natural numbers in terms of singletons.

type SomeSymbol = SomeSing (KindParam :: KindIs Symbol)

A definition of strings in terms of singletons.

Working with singletons

withSing :: SingI a => (Sing a -> b) -> b

A convenience function useful when we need to name a singleton value multiple times. Without this function, each use of sing could potentially refer to a different singleton, and one has to use type signatures to ensure that they are the same.

singThat :: SingRep a => (Demote a -> Bool) -> Maybe (Sing a)

A convenience function that names a singleton satisfying a certain property. If the singleton does not satisfy the property, then the function returns Nothing. The property is expressed in terms of the underlying representation of the singleton.

Functions on type nats

class (m <=? n) ~ True => m (<=) n

Comparison of type-level naturals.

type family m (<=?) n :: Bool

type family m (+) n :: Nat

Addition of type-level naturals.

type family m (*) n :: Nat

Multiplication of type-level naturals.

type family m (^) n :: Nat

Exponentiation of type-level naturals.

type family m (-) n :: Nat

Subtraction of type-level naturals. Note that this operation is unspecified for some inputs.

Comparing for equality

data :~: where

A type that provides evidence for equality between two types.

Constructors

Refl :: a :~: a

eqSingNat :: Sing (m :: Nat) -> Sing (n :: Nat) -> Maybe (m :~: n)

Check if two type-natural singletons of potentially different types are indeed the same, by comparing their runtime representations.

eqSingSym :: Sing (m :: Symbol) -> Sing (n :: Symbol) -> Maybe (m :~: n)

Check if two symbol singletons of potentially different types are indeed the same, by comparing their runtime representations.

Destructing type-nat singletons.

isZero :: Sing n -> IsZero n

data IsZero where

Constructors

IsZero :: IsZero 0

IsSucc :: !(Sing n) -> IsZero (n + 1)

Matching on type-nats

data Nat1

Unary implementation of natural numbers. Used both at the type and at the value level.

Constructors

Zero

Succ Nat1

type family FromNat1 n :: Nat

Kind parameters

data KindIs a

(Kind) A kind useful for passing kinds as parameters.

Constructors

KindParam

type Demote a = DemoteRep (KindOf a)

A convenient name for the type used to representing the values for a particular singleton family. For example, Demote 2 ~ Integer, and also Demote 3 ~ Integer, but Demote Hello ~ String.

type KindOf a = (KindParam :: KindIs k)

A shortcut for naming the kind parameter corresponding to the kind of a some type. For example, KindOf Int ~ (KindParam :: KindIs *), but KindOf 2 ~ (KindParam :: KindIs Nat).