GHC.Weak.Finalize

Handling exceptions

When an exception is thrown by a finalizer called by the garbage collector, GHC calls a global handler which can be set with setFinalizerExceptionHandler. Note that any exceptions thrown by this handler will be ignored.

setFinalizerExceptionHandler :: (SomeException -> IO ()) -> IO ()

Set the global action called to report exceptions thrown by weak pointer finalizers to the user.

Since: base-4.18.0.0

getFinalizerExceptionHandler :: IO (SomeException -> IO ())

Get the global action called to report exceptions thrown by weak pointer finalizers to the user.

Since: base-4.18.0.0

Internal

runFinalizerBatch :: Int -> Array# (State# RealWorld -> State# RealWorld) -> IO ()

Run a batch of finalizers from the garbage collector. We're given an array of finalizers and the length of the array, and we just call each one in turn.