System.CPUTime

Description

The standard CPUTime library.

Documentation

getCPUTime :: IO Integer

Computation getCPUTime returns the number of picoseconds CPU time used by the current program. The precision of this result is implementation-dependent.

cpuTimePrecision :: Integer

The cpuTimePrecision constant is the smallest measurable difference in CPU time that the implementation can record, and is given as an integral number of picoseconds.