System.IO.Error

Description

Standard IO Errors.

I/O errors

type IOError = IOException

The Haskell 2010 type for exceptions in the IO monad. Any I/O operation may raise an IOException instead of returning a result. For a more general type of exception, including also those that arise in pure code, see Exception.

In Haskell 2010, this is an opaque type.

userError :: String -> IOError

Construct an IOException value with a string describing the error. The fail method of the IO instance of the Monad class raises a userError, thus:

instance Monad IO where
  ...
  fail s = ioError (userError s)

mkIOError :: IOErrorType -> String -> Maybe Handle -> Maybe FilePath -> IOError

Construct an IOException of the given type where the second argument describes the error location and the third and fourth argument contain the file handle and file path of the file involved in the error if applicable.

annotateIOError :: IOError -> String -> Maybe Handle -> Maybe FilePath -> IOError

Adds a location description and maybe a file path and file handle to an IOException. If any of the file handle or file path is not given the corresponding value in the IOException remains unaltered.

Classifying I/O errors

isAlreadyExistsError :: IOError -> Bool

An error indicating that an IO operation failed because one of its arguments already exists.

isDoesNotExistError :: IOError -> Bool

An error indicating that an IO operation failed because one of its arguments does not exist.

isAlreadyInUseError :: IOError -> Bool

An error indicating that an IO operation failed because one of its arguments is a single-use resource, which is already being used (for example, opening the same file twice for writing might give this error).

isFullError :: IOError -> Bool

An error indicating that an IO operation failed because the device is full.

isEOFError :: IOError -> Bool

An error indicating that an IO operation failed because the end of file has been reached.

isIllegalOperation :: IOError -> Bool

An error indicating that an IO operation failed because the operation was not possible. Any computation which returns an IO result may fail with isIllegalOperation. In some cases, an implementation will not be able to distinguish between the possible error causes. In this case it should fail with isIllegalOperation.

isPermissionError :: IOError -> Bool

An error indicating that an IO operation failed because the user does not have sufficient operating system privilege to perform that operation.

isUserError :: IOError -> Bool

A programmer-defined error value constructed using userError.

isResourceVanishedError :: IOError -> Bool

An error indicating that the operation failed because the resource vanished. See resourceVanishedErrorType.

Since: base-4.14.0.0

Attributes of I/O errors

ioeGetErrorType :: IOError -> IOErrorType

ioeGetLocation :: IOError -> String

ioeGetErrorString :: IOError -> String

ioeGetHandle :: IOError -> Maybe Handle

ioeGetFileName :: IOError -> Maybe FilePath

ioeSetErrorType :: IOError -> IOErrorType -> IOError

ioeSetErrorString :: IOError -> String -> IOError

ioeSetLocation :: IOError -> String -> IOError

ioeSetHandle :: IOError -> Handle -> IOError

ioeSetFileName :: IOError -> FilePath -> IOError

Types of I/O error

data IOErrorType

An abstract type that contains a value for each variant of IOException.

alreadyExistsErrorType :: IOErrorType

I/O error where the operation failed because one of its arguments already exists.

doesNotExistErrorType :: IOErrorType

I/O error where the operation failed because one of its arguments does not exist.

alreadyInUseErrorType :: IOErrorType

I/O error where the operation failed because one of its arguments is a single-use resource, which is already being used.

fullErrorType :: IOErrorType

I/O error where the operation failed because the device is full.

eofErrorType :: IOErrorType

I/O error where the operation failed because the end of file has been reached.

illegalOperationErrorType :: IOErrorType

I/O error where the operation is not possible.

permissionErrorType :: IOErrorType

I/O error where the operation failed because the user does not have sufficient operating system privilege to perform that operation.

userErrorType :: IOErrorType

I/O error that is programmer-defined.

resourceVanishedErrorType :: IOErrorType

I/O error where the operation failed because the resource vanished. This happens when, for example, attempting to write to a closed socket or attempting to write to a named pipe that was deleted.

Since: base-4.14.0.0

IOErrorType predicates

isAlreadyExistsErrorType :: IOErrorType -> Bool

I/O error where the operation failed because one of its arguments already exists.

isDoesNotExistErrorType :: IOErrorType -> Bool

I/O error where the operation failed because one of its arguments does not exist.

isAlreadyInUseErrorType :: IOErrorType -> Bool

I/O error where the operation failed because one of its arguments is a single-use resource, which is already being used.

isFullErrorType :: IOErrorType -> Bool

I/O error where the operation failed because the device is full.

isEOFErrorType :: IOErrorType -> Bool

I/O error where the operation failed because the end of file has been reached.

isIllegalOperationErrorType :: IOErrorType -> Bool

I/O error where the operation is not possible.

isPermissionErrorType :: IOErrorType -> Bool

I/O error where the operation failed because the user does not have sufficient operating system privilege to perform that operation.

isUserErrorType :: IOErrorType -> Bool

I/O error that is programmer-defined.

isResourceVanishedErrorType :: IOErrorType -> Bool

I/O error where the operation failed because the resource vanished. See resourceVanishedErrorType.

Since: base-4.14.0.0

Throwing and catching I/O errors

ioError :: IOError -> IO a

Raise an IOException in the IO monad.

catchIOError :: IO a -> (IOError -> IO a) -> IO a

The catchIOError function establishes a handler that receives any IOException raised in the action protected by catchIOError. An IOException is caught by the most recent handler established by one of the exception handling functions. These handlers are not selective: all IOExceptions are caught. Exception propagation must be explicitly provided in a handler by re-raising any unwanted exceptions. For example, in

f = catchIOError g (\e -> if IO.isEOFError e then return [] else ioError e)

the function f returns [] when an end-of-file exception (cf. isEOFError) occurs in g; otherwise, the exception is propagated to the next outer handler.

When an exception propagates outside the main program, the Haskell system prints the associated IOException value and exits the program.

Non-I/O exceptions are not caught by this variant; to catch all exceptions, use catch from Control.Exception.

Since: base-4.4.0.0

tryIOError :: IO a -> IO (Either IOError a)

The construct tryIOError comp exposes IO errors which occur within a computation, and which are not fully handled.

Non-I/O exceptions are not caught by this variant; to catch all exceptions, use try from Control.Exception.

Since: base-4.4.0.0

modifyIOError :: (IOError -> IOError) -> IO a -> IO a

Catch any IOException that occurs in the computation and throw a modified version.