System.Info

Description

Information about the characteristics of the host system lucky enough to run your program.

For a comprehensive listing of supported platforms, please refer to https://gitlab.haskell.org/ghc/ghc/-/wikis/platforms

Documentation

os :: String

The operating system on which the program is running. Common values include:

"darwin" — macOS

"freebsd"

"linux"

"linux-android"

"mingw32" — Windows

"netbsd"

"openbsd"

arch :: String

The machine architecture on which the program is running. Common values include:

"aarch64"

"alpha"

"arm"

"hppa"

"hppa1_1"

"i386"

"ia64"

"m68k"

"mips"

"mipseb"

"mipsel"

"nios2"

"powerpc"

"powerpc64"

"powerpc64le"

"riscv32"

"riscv64"

"loongarch32"

"loongarch64"

"rs6000"

"s390"

"s390x"

"sh4"

"sparc"

"sparc64"

"vax"

"x86_64"

compilerName :: String

The Haskell implementation with which the program was compiled or is being interpreted. On the GHC platform, the value is "ghc".

compilerVersion :: Version

The version of compilerName with which the program was compiled or is being interpreted.

Example

Expand

ghci> compilerVersion
Version {versionBranch = [8,8], versionTags = []}

fullCompilerVersion :: Version

The full version of compilerName with which the program was compiled or is being interpreted. It includes the major, minor, revision and an additional identifier, generally in the form "yearmonthday".