Type.Reflection

Description

This provides a type-indexed type representation mechanism, similar to that described by,

Simon Peyton-Jones, Stephanie Weirich, Richard Eisenberg, Dimitrios Vytiniotis. "A reflection on types". Proc. Philip Wadler's 60th birthday Festschrift, Edinburgh (April 2016).

The interface provides TypeRep, a type representation which can be safely decomposed and composed. See Data.Dynamic for an example of this.

Since: base-4.10.0.0

The Typeable class

class Typeable (a :: k)

The class Typeable allows a concrete representation of a type to be calculated.

Minimal complete definition

typeRep#

typeRep :: Typeable a => TypeRep a

withTypeable :: forall k (a :: k) rep (r :: TYPE rep). TypeRep a -> (Typeable a => r) -> r

Use a TypeRep as Typeable evidence.

The TypeRep pattern synonym brings a Typeable constraint into scope and can be used in place of withTypeable.

f :: TypeRep a -> ..
f rep = withTypeable {- Typeable a in scope -}

f :: TypeRep a -> ..
f TypeRep = {- Typeable a in scope -}

Propositional equality

data a :~: b where infix 4

Propositional equality. If a :~: b is inhabited by some terminating value, then the type a is the same as the type b. To use this equality in practice, pattern-match on the a :~: b to get out the Refl constructor; in the body of the pattern-match, the compiler knows that a ~ b.

Since: base-4.7.0.0

Constructors

Refl :: a :~: a

data a :~~: b where infix 4

Kind heterogeneous propositional equality. Like :~:, a :~~: b is inhabited by a terminating value if and only if a is the same type as b.

Since: base-4.10.0.0

Constructors

HRefl :: a :~~: a

Type representations

Type-Indexed

data TypeRep a

TypeRep is a concrete representation of a (monomorphic) type. TypeRep supports reasonably efficient equality. See Note [Grand plan for Typeable] in GHC.Tc.Instance.Typeable

pattern TypeRep :: forall {k :: Type} (a :: k). () => Typeable @k a => TypeRep @k a

A explicitly bidirectional pattern synonym to construct a concrete representation of a type.

As an expression: Constructs a singleton TypeRep a given a implicit 'Typeable a' constraint:

TypeRep @a :: Typeable a => TypeRep a

As a pattern: Matches on an explicit TypeRep a witness bringing an implicit Typeable a constraint into scope.

f :: TypeRep a -> ..
f TypeRep = {- Typeable a in scope -}

Since: base-4.17.0.0

typeOf :: Typeable a => a -> TypeRep a

pattern App :: forall k2 (t :: k2). () => forall k1 (a :: k1 -> k2) (b :: k1). t ~ a b => TypeRep a -> TypeRep b -> TypeRep t

A type application.

For instance,

typeRep @(Maybe Int) === App (typeRep @Maybe) (typeRep @Int)

Note that this will also match a function type,

typeRep @(Int# -> Char)
  ===
App (App arrow (typeRep @Int#)) (typeRep @Char)

where arrow :: TypeRep ((->) :: TYPE IntRep -> Type -> Type).

pattern Con :: forall k (a :: k). () => NotApplication a => TyCon -> TypeRep a

Pattern match on a type constructor

pattern Con' :: forall k (a :: k). () => NotApplication a => TyCon -> [SomeTypeRep] -> TypeRep a

Pattern match on a type constructor including its instantiated kind variables.

For instance,

App (Con' proxyTyCon ks) intRep = typeRep @(Proxy @Int)

will bring into scope,

proxyTyCon :: TyCon
ks         == [someTypeRep Type] :: [SomeTypeRep]
intRep     == typeRep Int

pattern Fun :: forall k (fun :: k). () => forall (r1 :: RuntimeRep) (r2 :: RuntimeRep) (arg :: TYPE r1) (res :: TYPE r2). (k ~ Type, fun ~~ (arg -> res)) => TypeRep arg -> TypeRep res -> TypeRep fun

The function type constructor.

For instance,

typeRep @(Int -> Char) === Fun (typeRep @Int) (typeRep @Char)

typeRepTyCon :: TypeRep a -> TyCon

Observe the type constructor of a type representation

rnfTypeRep :: TypeRep a -> ()

Helper to fully evaluate TypeRep for use as NFData(rnf) implementation

Since: base-4.8.0.0

eqTypeRep :: forall k1 k2 (a :: k1) (b :: k2). TypeRep a -> TypeRep b -> Maybe (a :~~: b)

Type equality

Since: base-4.10

typeRepKind :: TypeRep (a :: k) -> TypeRep k

Observe the kind of a type.

splitApps :: TypeRep a -> (TyCon, [SomeTypeRep])

Quantified

data SomeTypeRep where

A non-indexed type representation.

Constructors

SomeTypeRep :: forall k (a :: k). !(TypeRep a) -> SomeTypeRep

someTypeRep :: forall proxy a. Typeable a => proxy a -> SomeTypeRep

Takes a value of type a and returns a concrete representation of that type.

Since: base-4.7.0.0

someTypeRepTyCon :: SomeTypeRep -> TyCon

Observe the type constructor of a quantified type representation.

rnfSomeTypeRep :: SomeTypeRep -> ()

Helper to fully evaluate SomeTypeRep for use as NFData(rnf) implementation

Since: base-4.10.0.0

Type constructors

data TyCon

tyConPackage :: TyCon -> String

tyConModule :: TyCon -> String

tyConName :: TyCon -> String

rnfTyCon :: TyCon -> ()

Module names

data Module

moduleName :: Module -> String

modulePackage :: Module -> String

rnfModule :: Module -> ()

Helper to fully evaluate TyCon for use as NFData(rnf) implementation

Since: base-4.8.0.0