Type.Reflection.Unsafe

Description

The representations of the types TyCon and TypeRep, and the function mkTyCon which is used by derived instances of Typeable to construct TyCons.

Be warned, these functions can be used to construct ill-kinded type representations.

Type representations

data TypeRep a

TypeRep is a concrete representation of a (monomorphic) type. TypeRep supports reasonably efficient equality. See Note [Grand plan for Typeable] in GHC.Tc.Instance.Typeable

mkTrApp :: forall k1 k2 (a :: k1 -> k2) (b :: k1). TypeRep (a :: k1 -> k2) -> TypeRep (b :: k1) -> TypeRep (a b)

Construct a representation for a type application.

mkTyCon

Arguments

:: String
package name

-> String
module name

-> String
the name of the type constructor

-> Int
number of kind variables

-> KindRep
kind representation

-> TyCon
A unique TyCon object Used when the strings are dynamically allocated, eg from binary deserialisation

Exquisitely unsafe.

typeRepFingerprint :: TypeRep a -> Fingerprint

Observe the Fingerprint of a type representation

Since: base-4.8.0.0

someTypeRepFingerprint :: SomeTypeRep -> Fingerprint

Kind representations

data KindRep

The representation produced by GHC for conjuring up the kind of a TypeRep.

Constructors

KindRepTyConApp TyCon [KindRep]

KindRepVar !KindBndr

KindRepApp KindRep KindRep

KindRepFun KindRep KindRep

KindRepTYPE !RuntimeRep

KindRepTypeLitS TypeLitSort Addr#

KindRepTypeLitD TypeLitSort [Char]

Bundled Patterns

pattern KindRepTypeLit :: TypeLitSort -> String -> KindRep

data TypeLitSort

Constructors

TypeLitSymbol

TypeLitNat

TypeLitChar

Type constructors

data TyCon

mkTrCon :: forall k (a :: k). TyCon -> [SomeTypeRep] -> TypeRep a

Construct a representation for a type constructor applied at a monomorphic kind.

Note that this is unsafe as it allows you to construct ill-kinded types.

tyConKindRep :: TyCon -> KindRep

tyConKindArgs :: TyCon -> Int

tyConFingerprint :: TyCon -> Fingerprint