Data.Foldable1

Description

A class of non-empty data structures that can be folded to a summary value.

Since: base-4.18.0.0

Documentation

class Foldable t => Foldable1 t where

Non-empty data structures that can be folded.

Since: base-4.18.0.0

Minimal complete definition

foldMap1 | foldrMap1

Methods

fold1 :: Semigroup m => t m -> m

Combine the elements of a structure using a semigroup. @since 4.18.0.0

foldMap1 :: Semigroup m => (a -> m) -> t a -> m

Map each element of the structure to a semigroup, and combine the results.

>>> foldMap1 Sum (1 :| [2, 3, 4])
Sum {getSum = 10}

Since: base-4.18.0.0

foldMap1' :: Semigroup m => (a -> m) -> t a -> m

A variant of foldMap1 that is strict in the accumulator.

>>> foldMap1' Sum (1 :| [2, 3, 4])
Sum {getSum = 10}

Since: base-4.18.0.0

toNonEmpty :: t a -> NonEmpty a

List of elements of a structure, from left to right.

>>> toNonEmpty (Identity 2)
2 :| []

Since: base-4.18.0.0

maximum :: Ord a => t a -> a

The largest element of a non-empty structure.

>>> maximum (32 :| [64, 8, 128, 16])
128

Since: base-4.18.0.0

minimum :: Ord a => t a -> a

The least element of a non-empty structure.

>>> minimum (32 :| [64, 8, 128, 16])
8

Since: base-4.18.0.0

head :: t a -> a

The first element of a non-empty structure.

>>> head (1 :| [2, 3, 4])
1

Since: base-4.18.0.0

last :: t a -> a

The last element of a non-empty structure.

>>> last (1 :| [2, 3, 4])
4

Since: base-4.18.0.0

foldrMap1 :: (a -> b) -> (a -> b -> b) -> t a -> b

Generalized foldr1. @since 4.18.0.0

foldlMap1' :: (a -> b) -> (b -> a -> b) -> t a -> b

Generalized foldl1'. @since 4.18.0.0

foldlMap1 :: (a -> b) -> (b -> a -> b) -> t a -> b

Generalized foldl1. @since 4.18.0.0

foldrMap1' :: (a -> b) -> (a -> b -> b) -> t a -> b

Generalized foldr1'. @since 4.18.0.0

foldr1 :: Foldable1 t => (a -> a -> a) -> t a -> a

Right-associative fold of a structure.

In the case of lists, foldr1, when applied to a binary operator, and a list, reduces the list using the binary operator, from right to left:

foldr1 f [x1, x2, ..., xn] == x1 `f` (x2 `f` ... (xn1 `f` xn )...)

Note that, since the head of the resulting expression is produced by an application of the operator to the first element of the list, foldr1 can produce a terminating expression from an infinite list.

For a general Foldable1 structure this should be semantically identical to,

foldr1 f = foldr1 f . toNonEmpty

Since: base-4.18.0.0

foldr1' :: Foldable1 t => (a -> a -> a) -> t a -> a

Right-associative fold of a structure, but with strict application of the operator.

Since: base-4.18.0.0

foldl1 :: Foldable1 t => (a -> a -> a) -> t a -> a

Left-associative fold of a structure.

In the case of lists, foldl1, when applied to a binary operator, and a list, reduces the list using the binary operator, from left to right:

foldl1 f [x1, x2, ..., xn] == (...((x1 `f` x2) `f`...) `f` xn

Note that to produce the outermost application of the operator the entire input list must be traversed. This means that foldl1 will diverge if given an infinite list.

Also note that if you want an efficient left-fold, you probably want to use foldl1' instead of foldl1. The reason for this is that latter does not force the "inner" results (e.g. x1 `f` x2 in the above example) before applying them to the operator (e.g. to (`f` x3)). This results in a thunk chain \(\mathcal{O}(n)\) elements long, which then must be evaluated from the outside-in.

For a general Foldable1 structure this should be semantically identical to,

foldl1 f z = foldl1 f . toNonEmpty

Since: base-4.18.0.0

foldl1' :: Foldable1 t => (a -> a -> a) -> t a -> a

Left-associative fold of a structure but with strict application of the operator.

This ensures that each step of the fold is forced to weak head normal form before being applied, avoiding the collection of thunks that would otherwise occur. This is often what you want to strictly reduce a finite list to a single, monolithic result (e.g. length).

For a general Foldable1 structure this should be semantically identical to,

foldl1' f z = foldl1 f . toNonEmpty

Since: base-4.18.0.0

intercalate1 :: (Foldable1 t, Semigroup m) => m -> t m -> m

Insert an m between each pair of t m.

>>> intercalate1 ", " $ "hello" :| ["how", "are", "you"]
"hello, how, are, you"

>>> intercalate1 ", " $ "hello" :| []
"hello"

>>> intercalate1 mempty $ "I" :| ["Am", "Fine", "You?"]
"IAmFineYou?"

Since: base-4.18.0.0

foldrM1 :: (Foldable1 t, Monad m) => (a -> a -> m a) -> t a -> m a

Monadic fold over the elements of a non-empty structure, associating to the right, i.e. from right to left.

Since: base-4.18.0.0

foldlM1 :: (Foldable1 t, Monad m) => (a -> a -> m a) -> t a -> m a

Monadic fold over the elements of a non-empty structure, associating to the left, i.e. from left to right.

Since: base-4.18.0.0

foldrMapM1 :: (Foldable1 t, Monad m) => (a -> m b) -> (a -> b -> m b) -> t a -> m b

Map variant of foldrM1.

Since: base-4.18.0.0

foldlMapM1 :: (Foldable1 t, Monad m) => (a -> m b) -> (b -> a -> m b) -> t a -> m b

Map variant of foldlM1.

Since: base-4.18.0.0

maximumBy :: Foldable1 t => (a -> a -> Ordering) -> t a -> a

The largest element of a non-empty structure with respect to the given comparison function.

Since: base-4.18.0.0

minimumBy :: Foldable1 t => (a -> a -> Ordering) -> t a -> a

The least element of a non-empty structure with respect to the given comparison function.

Since: base-4.18.0.0