Data.Functor.Compose

Description

Composition of functors.

Since: base-4.9.0.0

Documentation

newtype Compose f g a infixr 9

Right-to-left composition of functors. The composition of applicative functors is always applicative, but the composition of monads is not always a monad.

Constructors

Compose infixr 9

Fields

getCompose :: f (g a)