Data.Functor.Identity

Description

The identity functor and monad.

This trivial type constructor serves two purposes:

It can be used with functions parameterized by functor or monad classes.

It can be used as a base monad to which a series of monad transformers may be applied to construct a composite monad. Most monad transformer modules include the special case of applying the transformer to Identity. For example, State s is an abbreviation for StateT s Identity.

Since: 4.8.0.0

Documentation

newtype Identity a

Identity functor and monad. (a non-strict monad)

Since: 4.8.0.0

Constructors

Identity

Fields

runIdentity :: a