Data.Monoid

Description

A class for monoids (types with an associative binary operation that has an identity) with various general-purpose instances.

Monoid typeclass

class Monoid a where

The class of monoids (types with an associative binary operation that has an identity). Instances should satisfy the following laws:

mappend mempty x = x

mappend x mempty = x

mappend x (mappend y z) = mappend (mappend x y) z

mconcat = foldr mappend mempty

The method names refer to the monoid of lists under concatenation, but there are many other instances.

Minimal complete definition: mempty and mappend.

Some types can be viewed as a monoid in more than one way, e.g. both addition and multiplication on numbers. In such cases we often define newtypes and make those instances of Monoid, e.g. Sum and Product.

Methods

mempty :: a

Identity of mappend

mappend :: a -> a -> a

An associative operation

mconcat :: [a] -> a

Fold a list using the monoid. For most types, the default definition for mconcat will be used, but the function is included in the class definition so that an optimized version can be provided for specific types.

(<>) :: Monoid m => m -> m -> m

An infix synonym for mappend.

newtype Dual a

The dual of a monoid, obtained by swapping the arguments of mappend.

Constructors

Dual

Fields

getDual :: a

newtype Endo a

The monoid of endomorphisms under composition.

Constructors

Endo

Fields

appEndo :: a -> a

Bool wrappers

newtype All

Boolean monoid under conjunction.

Constructors

All

Fields

getAll :: Bool

newtype Any

Boolean monoid under disjunction.

Constructors

Any

Fields

getAny :: Bool

Num wrappers

newtype Sum a

Monoid under addition.

Constructors

Sum

Fields

getSum :: a

newtype Product a

Monoid under multiplication.

Constructors

Product

Fields

getProduct :: a

Maybe wrappers

To implement find or findLast on any Foldable:

 findLast :: Foldable t => (a -> Bool) -> t a -> Maybe a
 findLast pred = getLast . foldMap (x -> if pred x
                                            then Last (Just x)
                                            else Last Nothing)

Much of Data.Map's interface can be implemented with Data.Map.alter. Some of the rest can be implemented with a new alterA function and either First or Last:

 alterA :: (Applicative f, Ord k) =>
           (Maybe a -> f (Maybe a)) -> k -> Map k a -> f (Map k a)

 instance Monoid a => Applicative ((,) a)  -- from Control.Applicative

 insertLookupWithKey :: Ord k => (k -> v -> v -> v) -> k -> v
                     -> Map k v -> (Maybe v, Map k v)
 insertLookupWithKey combine key value =
   Arrow.first getFirst . alterA doChange key
   where
   doChange Nothing = (First Nothing, Just value)
   doChange (Just oldValue) =
     (First (Just oldValue),
      Just (combine key value oldValue))

newtype First a

Maybe monoid returning the leftmost non-Nothing value.

Constructors

First

Fields

getFirst :: Maybe a

newtype Last a

Maybe monoid returning the rightmost non-Nothing value.

Constructors

Last

Fields

getLast :: Maybe a