Data.Semigroup

Description

A type a is a Semigroup if it provides an associative function (<>) that lets you combine any two values of type a into one. Where being associative means that the following must always hold:

(a <> b) <> c == a <> (b <> c)

Examples

Expand

The Min Semigroup instance for Int is defined to always pick the smaller number:

>>> Min 1 <> Min 2 <> Min 3 <> Min 4 :: Min Int
Min {getMin = 1}

If we need to combine multiple values we can use the sconcat function to do so. We need to ensure however that we have at least one value to operate on, since otherwise our result would be undefined. It is for this reason that sconcat uses Data.List.NonEmpty.NonEmpty - a list that can never be empty:

>>> (1 :| [])
1 :| []               -- equivalent to [1] but guaranteed to be non-empty.

>>> (1 :| [2, 3, 4])
1 :| [2,3,4]          -- equivalent to [1,2,3,4] but guaranteed to be non-empty.

Equipped with this guaranteed to be non-empty data structure, we can combine values using sconcat and a Semigroup of our choosing. We can try the Min and Max instances of Int which pick the smallest, or largest number respectively:

>>> sconcat (1 :| [2, 3, 4]) :: Min Int
Min {getMin = 1}

>>> sconcat (1 :| [2, 3, 4]) :: Max Int
Max {getMax = 4}

String concatenation is another example of a Semigroup instance:

>>> "foo" <> "bar"
"foobar"

A Semigroup is a generalization of a Monoid. Yet unlike the Semigroup, the Monoid requires the presence of a neutral element (mempty) in addition to the associative operator. The requirement for a neutral element prevents many types from being a full Monoid, like Data.List.NonEmpty.NonEmpty.

Note that the use of (<>) in this module conflicts with an operator with the same name that is being exported by Data.Monoid. However, this package re-exports (most of) the contents of Data.Monoid, so to use semigroups and monoids in the same package just

import Data.Semigroup

Since: base-4.9.0.0

Documentation

class Semigroup a where

The class of semigroups (types with an associative binary operation).

Instances should satisfy the following:

Associativity

x <> (y <> z) = (x <> y) <> z

You can alternatively define sconcat instead of (<>), in which case the laws are:

Unit

sconcat (pure x) = x

Multiplication

sconcat (join xss) = sconcat (fmap sconcat xss)

Since: base-4.9.0.0

Minimal complete definition

(<>) | sconcat

Methods

(<>) :: a -> a -> a infixr 6

An associative operation.

Examples

Expand

>>> [1,2,3] <> [4,5,6]
[1,2,3,4,5,6]

>>> Just [1, 2, 3] <> Just [4, 5, 6]
Just [1,2,3,4,5,6]

>>> putStr "Hello, " <> putStrLn "World!"
Hello, World!

sconcat :: NonEmpty a -> a

Reduce a non-empty list with <>

The default definition should be sufficient, but this can be overridden for efficiency.

Examples

Expand

For the following examples, we will assume that we have:

>>> import Data.List.NonEmpty (NonEmpty (..))

>>> sconcat $ "Hello" :| [" ", "Haskell", "!"]
"Hello Haskell!"

>>> sconcat $ Just [1, 2, 3] :| [Nothing, Just [4, 5, 6]]
Just [1,2,3,4,5,6]

>>> sconcat $ Left 1 :| [Right 2, Left 3, Right 4]
Right 2

stimes :: Integral b => b -> a -> a

Repeat a value n times.

The default definition will raise an exception for a multiplier that is <= 0. This may be overridden with an implementation that is total. For monoids it is preferred to use stimesMonoid.

By making this a member of the class, idempotent semigroups and monoids can upgrade this to execute in \(\mathcal{O}(1)\) by picking stimes = stimesIdempotent or stimes = stimesIdempotentMonoid respectively.

Examples

Expand

>>> stimes 4 [1]
[1,1,1,1]

>>> stimes 5 (putStr "hi!")
hi!hi!hi!hi!hi!

>>> stimes 3 (Right ":)")
Right ":)"

stimesMonoid :: (Integral b, Monoid a) => b -> a -> a

This is a valid definition of stimes for a Monoid.

Unlike the default definition of stimes, it is defined for 0 and so it should be preferred where possible.

stimesIdempotent :: Integral b => b -> a -> a

This is a valid definition of stimes for an idempotent Semigroup.

When x <> x = x, this definition should be preferred, because it works in \(\mathcal{O}(1)\) rather than \(\mathcal{O}(\log n)\).

stimesIdempotentMonoid :: (Integral b, Monoid a) => b -> a -> a

This is a valid definition of stimes for an idempotent Monoid.

When x <> x = x, this definition should be preferred, because it works in \(\mathcal{O}(1)\) rather than \(\mathcal{O}(\log n)\)

mtimesDefault :: (Integral b, Monoid a) => b -> a -> a

Repeat a value n times.

mtimesDefault n a = a <> a <> ... <> a  -- using <> (n-1) times

In many cases, stimes 0 a for a Monoid will produce mempty. However, there are situations when it cannot do so. In particular, the following situation is fairly common:

data T a = ...

class Constraint1 a
class Constraint1 a => Constraint2 a

instance Constraint1 a => Semigroup (T a)
instance Constraint2 a => Monoid (T a)

Since Constraint1 is insufficient to implement mempty, stimes for T a cannot do so.

When working with such a type, or when working polymorphically with Semigroup instances, mtimesDefault should be used when the multiplier might be zero. It is implemented using stimes when the multiplier is nonzero and mempty when it is zero.

Examples

Expand

>>> mtimesDefault 0 "bark"
[]

>>> mtimesDefault 3 "meow"
"meowmeowmeow"

Semigroups

newtype Min a

The Min Monoid and Semigroup always choose the smaller element as by the Ord instance and min of the contained type.

Examples

Expand

>>> Min 42 <> Min 3
Min 3

>>> sconcat $ Min 1 :| [ Min n | n <- [2 .. 100]]
Min {getMin = 1}

Constructors

Min

Fields

getMin :: a

newtype Max a

The Max Monoid and Semigroup always choose the bigger element as by the Ord instance and max of the contained type.

Examples

Expand

>>> Max 42 <> Max 3
Max 42

>>> sconcat $ Max 1 :| [ Max n | n <- [2 .. 100]]
Max {getMax = 100}

Constructors

Max

Fields

getMax :: a

newtype First a

Beware that Data.Semigroup.First is different from Data.Monoid.First. The former simply returns the first value, so Data.Semigroup.First Nothing <> x = Data.Semigroup.First Nothing. The latter returns the first non-Nothing, thus Data.Monoid.First Nothing <> x = x.

Examples

Expand

>>> First 0 <> First 10
First 0

>>> sconcat $ First 1 :| [ First n | n <- [2 ..] ]
First 1

Constructors

First

Fields

getFirst :: a

newtype Last a

Beware that Data.Semigroup.Last is different from Data.Monoid.Last. The former simply returns the last value, so x <> Data.Semigroup.Last Nothing = Data.Semigroup.Last Nothing. The latter returns the last non-Nothing, thus x <> Data.Monoid.Last Nothing = x.

Examples

Expand

>>> Last 0 <> Last 10
Last {getLast = 10}

>>> sconcat $ Last 1 :| [ Last n | n <- [2..]]
Last {getLast = * hangs forever *

Constructors

Last

Fields

getLast :: a

newtype WrappedMonoid m

Provide a Semigroup for an arbitrary Monoid.

NOTE: This is not needed anymore since Semigroup became a superclass of Monoid in base-4.11 and this newtype be deprecated at some point in the future.

Constructors

WrapMonoid

Fields

unwrapMonoid :: m

Re-exported monoids from Data.Monoid

newtype Dual a

The dual of a Monoid, obtained by swapping the arguments of mappend. | The dual of a Monoid, obtained by swapping the arguments of (<>).

Dual a <> Dual b == Dual (b <> a)

Examples

Expand

>>> Dual "Hello" <> Dual "World"
Dual {getDual = "WorldHello"}

>>> Dual (Dual "Hello") <> Dual (Dual "World")
Dual {getDual = Dual {getDual = "HelloWorld"}}

Constructors

Dual

Fields

getDual :: a

newtype Endo a

The monoid of endomorphisms under composition.

Endo f <> Endo g == Endo (f . g)

Examples

Expand

>>> let computation = Endo ("Hello, " ++) <> Endo (++ "!")
>>> appEndo computation "Haskell"
"Hello, Haskell!"

>>> let computation = Endo (*3) <> Endo (+1)
>>> appEndo computation 1
6

Constructors

Endo

Fields

appEndo :: a -> a

newtype All

Boolean monoid under conjunction (&&).

All x <> All y = All (x && y)

Examples

Expand

>>> All True <> mempty <> All False)
All {getAll = False}

>>> mconcat (map (\x -> All (even x)) [2,4,6,7,8])
All {getAll = False}

>>> All True <> mempty
All {getAll = True}

Constructors

All

Fields

getAll :: Bool

newtype Any

Boolean monoid under disjunction (||).

Any x <> Any y = Any (x || y)

Examples

Expand

>>> Any True <> mempty <> Any False
Any {getAny = True}

>>> mconcat (map (\x -> Any (even x)) [2,4,6,7,8])
Any {getAny = True}

>>> Any False <> mempty
Any {getAny = False}

Constructors

Any

Fields

getAny :: Bool

newtype Sum a

Monoid under addition.

Sum a <> Sum b = Sum (a + b)

Examples

Expand

>>> Sum 1 <> Sum 2 <> mempty
Sum {getSum = 3}

>>> mconcat [ Sum n | n <- [3 .. 9]]
Sum {getSum = 42}

Constructors

Sum

Fields

getSum :: a

newtype Product a

Monoid under multiplication.

Product x <> Product y == Product (x * y)

Examples

Expand

>>> Product 3 <> Product 4 <> mempty
Product {getProduct = 12}

>>> mconcat [ Product n | n <- [2 .. 10]]
Product {getProduct = 3628800}

Constructors

Product

Fields

getProduct :: a

Difference lists of a semigroup

diff :: Semigroup m => m -> Endo m

This lets you use a difference list of a Semigroup as a Monoid.

Examples

Expand

let hello = diff "Hello, "

>>> appEndo hello "World!"
"Hello, World!"

>>> appEndo (hello <> mempty) "World!"
"Hello, World!"

>>> appEndo (mempty <> hello) "World!"
"Hello, World!"

let world = diff "World"
let excl = diff "!"

>>> appEndo (hello <> (world <> excl)) mempty
"Hello, World!"

>>> appEndo ((hello <> world) <> excl) mempty
"Hello, World!"

cycle1 :: Semigroup m => m -> m

A generalization of cycle to an arbitrary Semigroup. May fail to terminate for some values in some semigroups.

Examples

Expand

>>> take 10 $ cycle1 [1, 2, 3]
[1,2,3,1,2,3,1,2,3,1]

>>> cycle1 (Right 1)
Right 1

>>> cycle1 (Left 1)
* hangs forever *

ArgMin, ArgMax

data Arg a b

Arg isn't itself a Semigroup in its own right, but it can be placed inside Min and Max to compute an arg min or arg max.

Examples

Expand

>>> minimum [ Arg (x * x) x | x <- [-10 .. 10] ]
Arg 0 0

>>> maximum [ Arg (-0.2*x^2 + 1.5*x + 1) x | x <- [-10 .. 10] ]
Arg 3.8 4.0

>>> minimum [ Arg (-0.2*x^2 + 1.5*x + 1) x | x <- [-10 .. 10] ]
Arg (-34.0) (-10.0)

Constructors

Arg

Fields

a

The argument used for comparisons in Eq and Ord.

b

The "value" exposed via the Functor, Foldable etc. instances.

type ArgMin a b = Min (Arg a b)

Examples

Expand

>>> Min (Arg 0 ()) <> Min (Arg 1 ())
Min {getMin = Arg 0 ()}

>>> minimum [ Arg (length name) name | name <- ["violencia", "lea", "pixie"]]
Arg 3 "lea"

type ArgMax a b = Max (Arg a b)

Examples

Expand

>>> Max (Arg 0 ()) <> Max (Arg 1 ())
Max {getMax = Arg 1 ()}

>>> maximum [ Arg (length name) name | name <- ["violencia", "lea", "pixie"]]
Arg 9 "violencia"