Data.Type.Equality

Description

Definition of propositional equality (:~:). Pattern-matching on a variable of type (a :~: b) produces a proof that a ~ b.

Since: 4.7.0.0

The equality type

data a :~: b where infix 4

Propositional equality. If a :~: b is inhabited by some terminating value, then the type a is the same as the type b. To use this equality in practice, pattern-match on the a :~: b to get out the Refl constructor; in the body of the pattern-match, the compiler knows that a ~ b.

Since: 4.7.0.0

Constructors

Refl :: a :~: a

Working with equality

sym :: (a :~: b) -> b :~: a

Symmetry of equality

trans :: (a :~: b) -> (b :~: c) -> a :~: c

Transitivity of equality

castWith :: (a :~: b) -> a -> b

Type-safe cast, using propositional equality

gcastWith :: (a :~: b) -> ((a ~ b) => r) -> r

Generalized form of type-safe cast using propositional equality

apply :: (f :~: g) -> (a :~: b) -> f a :~: g b

Apply one equality to another, respectively

inner :: (f a :~: g b) -> a :~: b

Extract equality of the arguments from an equality of a applied types

outer :: (f a :~: g b) -> f :~: g

Extract equality of type constructors from an equality of applied types

Inferring equality from other types

class TestEquality f where

This class contains types where you can learn the equality of two types from information contained in terms. Typically, only singleton types should inhabit this class.

Methods

testEquality :: f a -> f b -> Maybe (a :~: b)

Conditionally prove the equality of a and b.

Boolean type-level equality

type family a == b :: Bool infix 4

A type family to compute Boolean equality. Instances are provided only for open kinds, such as * and function kinds. Instances are also provided for datatypes exported from base. A poly-kinded instance is not provided, as a recursive definition for algebraic kinds is generally more useful.