Data.Typeable

Description

The Typeable class reifies types to some extent by associating type representations to types. These type representations can be compared, and one can in turn define a type-safe cast operation. To this end, an unsafe cast is guarded by a test for type (representation) equivalence. The module Data.Dynamic uses Typeable for an implementation of dynamics. The module Data.Data uses Typeable and type-safe cast (but not dynamics) to support the "Scrap your boilerplate" style of generic programming.

Compatibility Notes

Since GHC 8.2, GHC has supported type-indexed type representations. Data.Typeable provides type representations which are qualified over this index, providing an interface very similar to the Typeable notion seen in previous releases. For the type-indexed interface, see Type.Reflection.

Since GHC 7.8, Typeable is poly-kinded. The changes required for this might break some old programs involving Typeable. More details on this, including how to fix your code, can be found on the PolyTypeable wiki page

The Typeable class

class Typeable (a :: k)

The class Typeable allows a concrete representation of a type to be calculated.

Minimal complete definition

typeRep#

typeOf :: forall a. Typeable a => a -> TypeRep

Observe a type representation for the type of a value.

typeRep :: forall proxy a. Typeable a => proxy a -> TypeRep

Takes a value of type a and returns a concrete representation of that type.

Since: 4.7.0.0

Propositional equality

data a :~: b where infix 4

Propositional equality. If a :~: b is inhabited by some terminating value, then the type a is the same as the type b. To use this equality in practice, pattern-match on the a :~: b to get out the Refl constructor; in the body of the pattern-match, the compiler knows that a ~ b.

Since: 4.7.0.0

Constructors

Refl :: a :~: a

data (a :: k1) :~~: (b :: k2) where infix 4

Kind heterogeneous propositional equality. Like '(:~:)', a :~~: b is inhabited by a terminating value if and only if a is the same type as b.

Since: 4.10.0.0

Constructors

HRefl :: a :~~: a

Type-safe cast

cast :: forall a b. (Typeable a, Typeable b) => a -> Maybe b

The type-safe cast operation

eqT :: forall a b. (Typeable a, Typeable b) => Maybe (a :~: b)

Extract a witness of equality of two types

Since: 4.7.0.0

gcast :: forall a b c. (Typeable a, Typeable b) => c a -> Maybe (c b)

A flexible variation parameterised in a type constructor

Generalized casts for higher-order kinds

gcast1 :: forall c t t' a. (Typeable t, Typeable t') => c (t a) -> Maybe (c (t' a))

Cast over k1 -> k2

gcast2 :: forall c t t' a b. (Typeable t, Typeable t') => c (t a b) -> Maybe (c (t' a b))

Cast over k1 -> k2 -> k3

A canonical proxy type

data Proxy t

A concrete, poly-kinded proxy type

Constructors

Proxy

Type representations

type TypeRep = SomeTypeRep

A quantified type representation.

rnfTypeRep :: TypeRep -> ()

Force a TypeRep to normal form.

showsTypeRep :: TypeRep -> ShowS

Show a type representation

mkFunTy :: TypeRep -> TypeRep -> TypeRep

Build a function type.

Observing type representations

funResultTy :: TypeRep -> TypeRep -> Maybe TypeRep

Applies a type to a function type. Returns: Just u if the first argument represents a function of type t -> u and the second argument represents a function of type t. Otherwise, returns Nothing.

splitTyConApp :: TypeRep -> (TyCon, [TypeRep])

Splits a type constructor application. Note that if the type constructor is polymorphic, this will not return the kinds that were used.

typeRepArgs :: TypeRep -> [TypeRep]

Observe the argument types of a type representation

typeRepTyCon :: TypeRep -> TyCon

Observe the type constructor of a quantified type representation.

typeRepFingerprint :: TypeRep -> Fingerprint

Takes a value of type a and returns a concrete representation of that type.

Since: 4.7.0.0

Type constructors

data TyCon :: *

tyConPackage :: TyCon -> String

tyConModule :: TyCon -> String

tyConName :: TyCon -> String

rnfTyCon :: TyCon -> ()

tyConFingerprint :: TyCon -> Fingerprint

For backwards compatibility

typeOf1 :: forall t (a :: *). Typeable t => t a -> TypeRep

typeOf2 :: forall t (a :: *) (b :: *). Typeable t => t a b -> TypeRep

typeOf3 :: forall t (a :: *) (b :: *) (c :: *). Typeable t => t a b c -> TypeRep

typeOf4 :: forall t (a :: *) (b :: *) (c :: *) (d :: *). Typeable t => t a b c d -> TypeRep

typeOf5 :: forall t (a :: *) (b :: *) (c :: *) (d :: *) (e :: *). Typeable t => t a b c d e -> TypeRep

typeOf6 :: forall t (a :: *) (b :: *) (c :: *) (d :: *) (e :: *) (f :: *). Typeable t => t a b c d e f -> TypeRep

typeOf7 :: forall t (a :: *) (b :: *) (c :: *) (d :: *) (e :: *) (f :: *) (g :: *). Typeable t => t a b c d e f g -> TypeRep

type Typeable1 (a :: * -> *) = Typeable a

Deprecated: renamed to Typeable

type Typeable2 (a :: * -> * -> *) = Typeable a

Deprecated: renamed to Typeable

type Typeable3 (a :: * -> * -> * -> *) = Typeable a

Deprecated: renamed to Typeable

type Typeable4 (a :: * -> * -> * -> * -> *) = Typeable a

Deprecated: renamed to Typeable

type Typeable5 (a :: * -> * -> * -> * -> * -> *) = Typeable a

Deprecated: renamed to Typeable

type Typeable6 (a :: * -> * -> * -> * -> * -> * -> *) = Typeable a

Deprecated: renamed to Typeable

type Typeable7 (a :: * -> * -> * -> * -> * -> * -> * -> *) = Typeable a

Deprecated: renamed to Typeable