Data.Void

Description

A logically uninhabited data type, used to indicate that a given term should not exist.

Since: 4.8.0.0

Documentation

data Void

Uninhabited data type

Since: 4.8.0.0

absurd :: Void -> a

Since Void values logically don't exist, this witnesses the logical reasoning tool of "ex falso quodlibet".

Since: 4.8.0.0

vacuous :: Functor f => f Void -> f a

If Void is uninhabited then any Functor that holds only values of type Void is holding no values.

Since: 4.8.0.0