Data.Void

Description

A logically uninhabited data type, used to indicate that a given term should not exist.

Since: base-4.8.0.0

Documentation

data Void

Uninhabited data type

Since: base-4.8.0.0

absurd :: Void -> a

Since Void values logically don't exist, this witnesses the logical reasoning tool of "ex falso quodlibet".

>>> let x :: Either Void Int; x = Right 5
>>> :{
case x of
    Right r -> r
    Left l  -> absurd l
:}
5

Since: base-4.8.0.0

vacuous :: Functor f => f Void -> f a

If Void is uninhabited then any Functor that holds only values of type Void is holding no values. It is implemented in terms of fmap absurd.

Since: base-4.8.0.0