Foreign.C.Types

Description

Mapping of C types to corresponding Haskell types.

Representations of C types

These types are needed to accurately represent C function prototypes, in order to access C library interfaces in Haskell. The Haskell system is not required to represent those types exactly as C does, but the following guarantees are provided concerning a Haskell type CT representing a C type t:

If a C function prototype has t as an argument or result type, the use of CT in the corresponding position in a foreign declaration permits the Haskell program to access the full range of values encoded by the C type; and conversely, any Haskell value for CT has a valid representation in C.

sizeOf (undefined :: CT) will yield the same value as sizeof (t) in C.

alignment (undefined :: CT) matches the alignment constraint enforced by the C implementation for t.

The members peek and poke of the Storable class map all values of CT to the corresponding value of t and vice versa.

When an instance of Bounded is defined for CT, the values of minBound and maxBound coincide with t_MIN and t_MAX in C.

When an instance of Eq or Ord is defined for CT, the predicates defined by the type class implement the same relation as the corresponding predicate in C on t.

When an instance of Num, Read, Integral, Fractional, Floating, RealFrac, or RealFloat is defined for CT, the arithmetic operations defined by the type class implement the same function as the corresponding arithmetic operations (if available) in C on t.

When an instance of Bits is defined for CT, the bitwise operation defined by the type class implement the same function as the corresponding bitwise operation in C on t.

Integral types

These types are are represented as newtypes of types in Data.Int and Data.Word, and are instances of Eq, Ord, Num, Read, Show, Enum, Typeable, Storable, Bounded, Real, Integral and Bits.

newtype CChar

Haskell type representing the C char type.

Constructors

CChar Int8

newtype CSChar

Haskell type representing the C signed char type.

Constructors

CSChar Int8

newtype CUChar

Haskell type representing the C unsigned char type.

Constructors

CUChar Word8

newtype CShort

Haskell type representing the C short type.

Constructors

CShort Int16

newtype CUShort

Haskell type representing the C unsigned short type.

Constructors

CUShort Word16

newtype CInt

Haskell type representing the C int type.

Constructors

CInt Int32

newtype CUInt

Haskell type representing the C unsigned int type.

Constructors

CUInt Word32

newtype CLong

Haskell type representing the C long type.

Constructors

CLong Int64

newtype CULong

Haskell type representing the C unsigned long type.

Constructors

CULong Word64

newtype CPtrdiff

Haskell type representing the C ptrdiff_t type.

Constructors

CPtrdiff Int64

newtype CSize

Haskell type representing the C size_t type.

Constructors

CSize Word64

newtype CWchar

Haskell type representing the C wchar_t type.

Constructors

CWchar Int32

newtype CSigAtomic

Haskell type representing the C sig_atomic_t type.

Constructors

CSigAtomic Int32

newtype CLLong

Haskell type representing the C long long type.

Constructors

CLLong Int64

newtype CULLong

Haskell type representing the C unsigned long long type.

Constructors

CULLong Word64

newtype CIntPtr

Constructors

CIntPtr Int64

newtype CUIntPtr

Constructors

CUIntPtr Word64

newtype CIntMax

Constructors

CIntMax Int64

newtype CUIntMax

Constructors

CUIntMax Word64

Numeric types

These types are represented as newtypes of basic foreign types, and are instances of Eq, Ord, Num, Read, Show, Enum, Typeable and Storable.

newtype CClock

Haskell type representing the C clock_t type.

Constructors

CClock Word64

newtype CTime

Haskell type representing the C time_t type.

Constructors

CTime Int64

newtype CUSeconds

Haskell type representing the C useconds_t type.

Since: 4.4.0.0

Constructors

CUSeconds Word32

newtype CSUSeconds

Haskell type representing the C suseconds_t type.

Since: 4.4.0.0

Constructors

CSUSeconds Int32

To convert CTime to UTCTime, use the following:

\t -> posixSecondsToUTCTime (realToFrac t :: POSIXTime)

Floating types

These types are are represented as newtypes of Float and Double, and are instances of Eq, Ord, Num, Read, Show, Enum, Typeable, Storable, Real, Fractional, Floating, RealFrac and RealFloat.

newtype CFloat

Haskell type representing the C float type.

Constructors

CFloat Float

newtype CDouble

Haskell type representing the C double type.

Constructors

CDouble Double

Other types

data CFile

Haskell type representing the C FILE type.

data CFpos

Haskell type representing the C fpos_t type.

data CJmpBuf

Haskell type representing the C jmp_buf type.