GHC.Float

Description

The types Float and Double, the classes Floating and RealFloat and casting between Word32 and Float and Word64 and Double.

Documentation

class Fractional a => Floating a where

Trigonometric and hyperbolic functions and related functions.

The Haskell Report defines no laws for Floating. However, (+), (*) and exp are customarily expected to define an exponential field and have the following properties:

exp (a + b) = exp a * exp b

exp (fromInteger 0) = fromInteger 1

Minimal complete definition

pi, exp, log, sin, cos, asin, acos, atan, sinh, cosh, asinh, acosh, atanh

Methods

pi :: a

exp :: a -> a

log :: a -> a

sqrt :: a -> a

(**) :: a -> a -> a infixr 8

logBase :: a -> a -> a

sin :: a -> a

cos :: a -> a

tan :: a -> a

asin :: a -> a

acos :: a -> a

atan :: a -> a

sinh :: a -> a

cosh :: a -> a

tanh :: a -> a

asinh :: a -> a

acosh :: a -> a

atanh :: a -> a

log1p :: a -> a

log1p x computes log (1 + x), but provides more precise results for small (absolute) values of x if possible.

Since: base-4.9.0.0

expm1 :: a -> a

expm1 x computes exp x - 1, but provides more precise results for small (absolute) values of x if possible.

Since: base-4.9.0.0

log1pexp :: a -> a

log1pexp x computes log (1 + exp x), but provides more precise results if possible.

Examples:

if x is a large negative number, log (1 + exp x) will be imprecise for the reasons given in log1p.

if exp x is close to -1, log (1 + exp x) will be imprecise for the reasons given in expm1.

Since: base-4.9.0.0

log1mexp :: a -> a

log1mexp x computes log (1 - exp x), but provides more precise results if possible.

Examples:

if x is a large negative number, log (1 - exp x) will be imprecise for the reasons given in log1p.

if exp x is close to 1, log (1 - exp x) will be imprecise for the reasons given in expm1.

Since: base-4.9.0.0

class (RealFrac a, Floating a) => RealFloat a where

Efficient, machine-independent access to the components of a floating-point number.

Minimal complete definition

floatRadix, floatDigits, floatRange, decodeFloat, encodeFloat, isNaN, isInfinite, isDenormalized, isNegativeZero, isIEEE

Methods

floatRadix :: a -> Integer

a constant function, returning the radix of the representation (often 2)

floatDigits :: a -> Int

a constant function, returning the number of digits of floatRadix in the significand

floatRange :: a -> (Int, Int)

a constant function, returning the lowest and highest values the exponent may assume

decodeFloat :: a -> (Integer, Int)

The function decodeFloat applied to a real floating-point number returns the significand expressed as an Integer and an appropriately scaled exponent (an Int). If decodeFloat x yields (m,n), then x is equal in value to m*b^^n, where b is the floating-point radix, and furthermore, either m and n are both zero or else b^(d-1) <= abs m < b^d, where d is the value of floatDigits x. In particular, decodeFloat 0 = (0,0). If the type contains a negative zero, also decodeFloat (-0.0) = (0,0). The result of decodeFloat x is unspecified if either of isNaN x or isInfinite x is True.

encodeFloat :: Integer -> Int -> a

encodeFloat performs the inverse of decodeFloat in the sense that for finite x with the exception of -0.0, uncurry encodeFloat (decodeFloat x) = x. encodeFloat m n is one of the two closest representable floating-point numbers to m*b^^n (or ±Infinity if overflow occurs); usually the closer, but if m contains too many bits, the result may be rounded in the wrong direction.

exponent :: a -> Int

exponent corresponds to the second component of decodeFloat. exponent 0 = 0 and for finite nonzero x, exponent x = snd (decodeFloat x) + floatDigits x. If x is a finite floating-point number, it is equal in value to significand x * b ^^ exponent x, where b is the floating-point radix. The behaviour is unspecified on infinite or NaN values.

significand :: a -> a

The first component of decodeFloat, scaled to lie in the open interval (-1,1), either 0.0 or of absolute value >= 1/b, where b is the floating-point radix. The behaviour is unspecified on infinite or NaN values.

scaleFloat :: Int -> a -> a

multiplies a floating-point number by an integer power of the radix

isNaN :: a -> Bool

True if the argument is an IEEE "not-a-number" (NaN) value

isInfinite :: a -> Bool

True if the argument is an IEEE infinity or negative infinity

isDenormalized :: a -> Bool

True if the argument is too small to be represented in normalized format

isNegativeZero :: a -> Bool

True if the argument is an IEEE negative zero

isIEEE :: a -> Bool

True if the argument is an IEEE floating point number

atan2 :: a -> a -> a

a version of arctangent taking two real floating-point arguments. For real floating x and y, atan2 y x computes the angle (from the positive x-axis) of the vector from the origin to the point (x,y). atan2 y x returns a value in the range [-pi, pi]. It follows the Common Lisp semantics for the origin when signed zeroes are supported. atan2 y 1, with y in a type that is RealFloat, should return the same value as atan y. A default definition of atan2 is provided, but implementors can provide a more accurate implementation.

data FFFormat

Constructors

FFExponent

FFFixed

FFGeneric

clamp :: Int -> Int -> Int

Used to prevent exponent over/underflow when encoding floating point numbers. This is also the same as

\(x,y) -> max (-x) (min x y)

Example

Expand

>>> clamp (-10) 5
10

showFloat :: RealFloat a => a -> ShowS

Show a signed RealFloat value to full precision using standard decimal notation for arguments whose absolute value lies between 0.1 and 9,999,999, and scientific notation otherwise.

floatToDigits :: RealFloat a => Integer -> a -> ([Int], Int)

floatToDigits takes a base and a non-negative RealFloat number, and returns a list of digits and an exponent. In particular, if x>=0, and

floatToDigits base x = ([d1,d2,...,dn], e)

then

n >= 1

x = 0.d1d2...dn * (base**e)

0 <= di <= base-1

fromRat :: RealFloat a => Rational -> a

Converts a Rational value into any type in class RealFloat.

formatRealFloat :: RealFloat a => FFFormat -> Maybe Int -> a -> String

log1mexpOrd :: (Ord a, Floating a) => a -> a

Default implementation for log1mexp requiring Ord to test against a threshold to decide which implementation variant to use.

plusFloat :: Float -> Float -> Float

minusFloat :: Float -> Float -> Float

negateFloat :: Float -> Float

timesFloat :: Float -> Float -> Float

fabsFloat :: Float -> Float

integerToFloat# :: Integer -> Float#

Convert an Integer to a Float#

integerToBinaryFloat' :: RealFloat a => Integer -> a

Converts a positive integer to a floating-point value.

The value nearest to the argument will be returned. If there are two such values, the one with an even significand will be returned (i.e. IEEE roundTiesToEven).

The argument must be strictly positive, and floatRadix (undefined :: a) must be 2.

naturalToFloat# :: Natural -> Float#

Convert a Natural to a Float#

divideFloat :: Float -> Float -> Float

rationalToFloat :: Integer -> Integer -> Float

fromRat'' :: RealFloat a => Int -> Int -> Integer -> Integer -> a

properFractionFloat :: Integral b => Float -> (b, Float)

truncateFloat :: Integral b => Float -> b

roundFloat :: Integral b => Float -> b

floorFloat :: Integral b => Float -> b

ceilingFloat :: Integral b => Float -> b

expFloat :: Float -> Float

logFloat :: Float -> Float

sqrtFloat :: Float -> Float

sinFloat :: Float -> Float

cosFloat :: Float -> Float

tanFloat :: Float -> Float

asinFloat :: Float -> Float

acosFloat :: Float -> Float

atanFloat :: Float -> Float

sinhFloat :: Float -> Float

coshFloat :: Float -> Float

tanhFloat :: Float -> Float

powerFloat :: Float -> Float -> Float

asinhFloat :: Float -> Float

acoshFloat :: Float -> Float

atanhFloat :: Float -> Float

log1pFloat :: Float -> Float

expm1Float :: Float -> Float

isFloatFinite :: Float -> Int

isFloatNaN :: Float -> Int

isFloatInfinite :: Float -> Int

isFloatDenormalized :: Float -> Int

isFloatNegativeZero :: Float -> Int

showSignedFloat

Arguments

:: RealFloat a

=> (a -> ShowS)
a function that can show unsigned values

-> Int
the precedence of the enclosing context

-> a
the value to show

-> ShowS

plusDouble :: Double -> Double -> Double

minusDouble :: Double -> Double -> Double

negateDouble :: Double -> Double

timesDouble :: Double -> Double -> Double

fabsDouble :: Double -> Double

integerToDouble# :: Integer -> Double#

Convert an Integer to a Double#

naturalToDouble# :: Natural -> Double#

Encode a Natural (mantissa) into a Double#

divideDouble :: Double -> Double -> Double

rationalToDouble :: Integer -> Integer -> Double

expDouble :: Double -> Double

logDouble :: Double -> Double

sqrtDouble :: Double -> Double

sinDouble :: Double -> Double

cosDouble :: Double -> Double

tanDouble :: Double -> Double

asinDouble :: Double -> Double

acosDouble :: Double -> Double

atanDouble :: Double -> Double

sinhDouble :: Double -> Double

coshDouble :: Double -> Double

tanhDouble :: Double -> Double

powerDouble :: Double -> Double -> Double

asinhDouble :: Double -> Double

acoshDouble :: Double -> Double

atanhDouble :: Double -> Double

log1pDouble :: Double -> Double

expm1Double :: Double -> Double

properFractionDouble :: Integral b => Double -> (b, Double)

truncateDouble :: Integral b => Double -> b

roundDouble :: Integral b => Double -> b

ceilingDouble :: Integral b => Double -> b

floorDouble :: Integral b => Double -> b

isDoubleFinite :: Double -> Int

isDoubleNaN :: Double -> Int

isDoubleInfinite :: Double -> Int

isDoubleDenormalized :: Double -> Int

isDoubleNegativeZero :: Double -> Int

formatRealFloatAlt :: RealFloat a => FFFormat -> Maybe Int -> Bool -> a -> String

roundTo :: Int -> Int -> [Int] -> (Int, [Int])

expt :: Integer -> Int -> Integer

roundingMode# :: Integer -> Int# -> Int#

fromRat' :: RealFloat a => Rational -> a

minExpt :: Int

maxExpt :: Int

expts :: Array Int Integer

maxExpt10 :: Int

expts10 :: Array Int Integer

gtFloat :: Float -> Float -> Bool

geFloat :: Float -> Float -> Bool

ltFloat :: Float -> Float -> Bool

leFloat :: Float -> Float -> Bool

gtDouble :: Double -> Double -> Bool

geDouble :: Double -> Double -> Bool

leDouble :: Double -> Double -> Bool

ltDouble :: Double -> Double -> Bool

double2Float :: Double -> Float

float2Double :: Float -> Double

word2Double :: Word -> Double

word2Float :: Word -> Float

castWord32ToFloat :: Word32 -> Float

castWord32ToFloat w does a bit-for-bit copy from an integral value to a floating-point value.

Since: base-4.11.0.0

stgWord32ToFloat :: Word32# -> Float#

castFloatToWord32 :: Float -> Word32

castFloatToWord32 f does a bit-for-bit copy from a floating-point value to an integral value.

Since: base-4.11.0.0

stgFloatToWord32 :: Float# -> Word32#

castWord64ToDouble :: Word64 -> Double

castWord64ToDouble w does a bit-for-bit copy from an integral value to a floating-point value.

Since: base-4.11.0.0

stgWord64ToDouble :: Word64# -> Double#

castDoubleToWord64 :: Double -> Word64

castFloatToWord64 f does a bit-for-bit copy from a floating-point value to an integral value.

Since: base-4.11.0.0

stgDoubleToWord64 :: Double# -> Word64#

data Float

Single-precision floating point numbers. It is desirable that this type be at least equal in range and precision to the IEEE single-precision type.

Constructors

F# Float#

data Double

Double-precision floating point numbers. It is desirable that this type be at least equal in range and precision to the IEEE double-precision type.

Constructors

D# Double#

data Float# :: TYPE 'FloatRep

data Double# :: TYPE 'DoubleRep

double2Int :: Double -> Int

int2Double :: Int -> Double

float2Int :: Float -> Int

int2Float :: Int -> Float

Monomorphic equality operators

See GHC.Classes#matching_overloaded_methods_in_rules

eqFloat :: Float -> Float -> Bool

eqDouble :: Double -> Double -> Bool

Orphan instances

Enum Double
Since: base-2.1

Instance details

Methods

succ :: Double -> Double

pred :: Double -> Double

toEnum :: Int -> Double

fromEnum :: Double -> Int

enumFrom :: Double -> [Double]

enumFromThen :: Double -> Double -> [Double]

enumFromTo :: Double -> Double -> [Double]

enumFromThenTo :: Double -> Double -> Double -> [Double]

Enum Float
Since: base-2.1

Instance details

Methods

succ :: Float -> Float

pred :: Float -> Float

toEnum :: Int -> Float

fromEnum :: Float -> Int

enumFrom :: Float -> [Float]

enumFromThen :: Float -> Float -> [Float]

enumFromTo :: Float -> Float -> [Float]

enumFromThenTo :: Float -> Float -> Float -> [Float]

Num Double
Note that due to the presence of NaN, not all elements of Double have an additive inverse.

>>> 0/0 + (negate 0/0 :: Double)
NaN

Also note that due to the presence of -0, Double's Num instance doesn't have an additive identity

>>> 0 + (-0 :: Double)
0.0

Since: base-2.1

Instance details

Methods

(+) :: Double -> Double -> Double

(-) :: Double -> Double -> Double

(*) :: Double -> Double -> Double

negate :: Double -> Double

abs :: Double -> Double

signum :: Double -> Double

fromInteger :: Integer -> Double

Num Float
Note that due to the presence of NaN, not all elements of Float have an additive inverse.

>>> 0/0 + (negate 0/0 :: Float)
NaN

Also note that due to the presence of -0, Float's Num instance doesn't have an additive identity

>>> 0 + (-0 :: Float)
0.0

Since: base-2.1

Instance details

Methods

(+) :: Float -> Float -> Float

(-) :: Float -> Float -> Float

(*) :: Float -> Float -> Float

negate :: Float -> Float

abs :: Float -> Float

signum :: Float -> Float

fromInteger :: Integer -> Float

Fractional Double
Note that due to the presence of NaN, not all elements of Double have an multiplicative inverse.

>>> 0/0 * (recip 0/0 :: Double)
NaN

Additionally, because of NaN this instance does not obey the left-inverse law for toRational/fromRational.

Since: base-2.1

Instance details

Methods

(/) :: Double -> Double -> Double

recip :: Double -> Double

fromRational :: Rational -> Double

Fractional Float
Note that due to the presence of NaN, not all elements of Float have an multiplicative inverse.

>>> 0/0 * (recip 0/0 :: Float)
NaN

Additionally, because of NaN this instance does not obey the left-inverse law for toRational/fromRational.

Since: base-2.1

Instance details

Methods

(/) :: Float -> Float -> Float

recip :: Float -> Float

fromRational :: Rational -> Float

Real Double
Since: base-2.1

Instance details

Methods

toRational :: Double -> Rational

Real Float
Since: base-2.1

Instance details

Methods

toRational :: Float -> Rational

RealFrac Double
Since: base-2.1

Instance details

Methods

properFraction :: Integral b => Double -> (b, Double)

truncate :: Integral b => Double -> b

round :: Integral b => Double -> b

ceiling :: Integral b => Double -> b

floor :: Integral b => Double -> b

RealFrac Float
Since: base-2.1

Instance details

Methods

properFraction :: Integral b => Float -> (b, Float)

truncate :: Integral b => Float -> b

round :: Integral b => Float -> b

ceiling :: Integral b => Float -> b

floor :: Integral b => Float -> b

Show Double
Since: base-2.1

Instance details

Methods

showsPrec :: Int -> Double -> ShowS

show :: Double -> String

showList :: [Double] -> ShowS

Show Float
Since: base-2.1

Instance details

Methods

showsPrec :: Int -> Float -> ShowS

show :: Float -> String

showList :: [Float] -> ShowS