GHC.IO.Encoding.CodePage.API

Documentation

mkCodePageEncoding :: CodingFailureMode -> Word32 -> TextEncoding

Returns a slow but correct implementation of TextEncoding using the Win32 API.

This is useful for supporting DBCS text encoding on the console without having to statically link in huge code tables into all of our executables, or just as a fallback mechanism if a new code page is introduced that we don't know how to deal with ourselves yet.