GHC.IO.SubSystem

Description

The SubSystem control interface. These methods can be used to disambiguate between the two operations.

Documentation

withIoSubSystem :: (IoSubSystem -> IO a) -> IO a

withIoSubSystem' :: (IoSubSystem -> a) -> a

whenIoSubSystem :: IoSubSystem -> IO () -> IO ()

ioSubSystem :: IoSubSystem

data IoSubSystem

The I/O SubSystem to use in the program.

Since: base-4.9.0.0

Constructors

IoPOSIX
Use a POSIX I/O Sub-System

IoNative
Use platform native Sub-System. For unix OSes this is the same as IoPOSIX, but on Windows this means use the Windows native APIs for I/O, including IOCP and RIO.

conditional :: a -> a -> a

Conditionally execute an action depending on the configured I/O subsystem. On POSIX systems always execute the first action. On windows execute the second action if WINIO as active, otherwise fall back to the first action.

(<!>) :: a -> a -> a infixl 7

Infix version of conditional. posix ! windows == conditional posix windows

isWindowsNativeIO :: Bool