GHC.Natural

Description

The arbitrary-precision Natural number type.

Note: This is an internal GHC module with an API subject to change. It's recommended use the Numeric.Natural module to import the Natural type.

Since: 4.8.0.0

The Natural number type

Warning: The internal implementation of Natural (i.e. which constructors are available) depends on the Integer backend used!

data Natural

Type representing arbitrary-precision non-negative integers.

Operations whose result would be negative throw (Underflow :: ArithException).

Since: 4.8.0.0

Constructors

NatS# GmpLimb#
in [0, maxBound::Word]

NatJ# !BigNat
in ]maxBound::Word, +inf[

Invariant: NatJ# is used iff value doesn't fit in NatS# constructor.

isValidNatural :: Natural -> Bool

Test whether all internal invariants are satisfied by Natural value

This operation is mostly useful for test-suites and/or code which constructs Integer values directly.

Since: 4.8.0.0

Conversions

wordToNatural :: Word -> Natural

Construct Natural from Word value.

Since: 4.8.0.0

naturalToWordMaybe :: Natural -> Maybe Word

Try downcasting Natural to Word value. Returns Nothing if value doesn't fit in Word.

Since: 4.8.0.0

Checked subtraction

minusNaturalMaybe :: Natural -> Natural -> Maybe Natural

Natural subtraction. Returns Nothings for non-positive results.

Since: 4.8.0.0

Modular arithmetic

powModNatural :: Natural -> Natural -> Natural -> Natural

"powModNatural b e m" computes base b raised to exponent e modulo m.

Since: 4.8.0.0