GHC.Num

Description

The Num class and the Integer type.

Documentation

class Num a where

Basic numeric class.

The Haskell Report defines no laws for Num. However, (+) and (*) are customarily expected to define a ring and have the following properties:

Associativity of (+)

(x + y) + z = x + (y + z)

Commutativity of (+)

x + y = y + x

fromInteger 0 is the additive identity

x + fromInteger 0 = x

negate gives the additive inverse

x + negate x = fromInteger 0

Associativity of (*)

(x * y) * z = x * (y * z)

fromInteger 1 is the multiplicative identity

x * fromInteger 1 = x and fromInteger 1 * x = x

Distributivity of (*) with respect to (+)

a * (b + c) = (a * b) + (a * c) and (b + c) * a = (b * a) + (c * a)

Coherence with toInteger

if the type also implements Integral, then fromInteger is a left inverse for toInteger, i.e. fromInteger (toInteger i) == i

Note that it isn't customarily expected that a type instance of both Num and Ord implement an ordered ring. Indeed, in base only Integer and Rational do.

Minimal complete definition

(+), (*), abs, signum, fromInteger, (negate | (-))

Methods

(+) :: a -> a -> a infixl 6

(-) :: a -> a -> a infixl 6

(*) :: a -> a -> a infixl 7

negate :: a -> a

Unary negation.

abs :: a -> a

Absolute value.

signum :: a -> a

Sign of a number. The functions abs and signum should satisfy the law:

abs x * signum x == x

For real numbers, the signum is either -1 (negative), 0 (zero) or 1 (positive).

fromInteger :: Integer -> a

Conversion from an Integer. An integer literal represents the application of the function fromInteger to the appropriate value of type Integer, so such literals have type (Num a) => a.

quotRemInteger :: Integer -> Integer -> (# Integer, Integer #)

Deprecated: Use integerQuotRem# instead

subtract :: Num a => a -> a -> a

the same as flip (-).

Because - is treated specially in the Haskell grammar, (- e) is not a section, but an application of prefix negation. However, (subtract exp) is equivalent to the disallowed section.

module GHC.Num.Integer

module GHC.Num.Natural

module GHC.Natural

module GHC.Integer