GHC.Real

Description

The types Ratio and Rational, and the classes Real, Fractional, Integral, and RealFrac.

Documentation

divZeroError :: a

ratioZeroDenominatorError :: a

overflowError :: a

underflowError :: a

data Ratio a

Rational numbers, with numerator and denominator of some Integral type.

Note that Ratio's instances inherit the deficiencies from the type parameter's. For example, Ratio Natural's Num instance has similar problems to Natural's.

Constructors

!a :% !a

type Rational = Ratio Integer

Arbitrary-precision rational numbers, represented as a ratio of two Integer values. A rational number may be constructed using the % operator.

ratioPrec :: Int

ratioPrec1 :: Int

infinity :: Rational

notANumber :: Rational

(%) :: Integral a => a -> a -> Ratio a infixl 7

Forms the ratio of two integral numbers.

numerator :: Ratio a -> a

Extract the numerator of the ratio in reduced form: the numerator and denominator have no common factor and the denominator is positive.

denominator :: Ratio a -> a

Extract the denominator of the ratio in reduced form: the numerator and denominator have no common factor and the denominator is positive.

reduce :: Integral a => a -> a -> Ratio a

reduce is a subsidiary function used only in this module. It normalises a ratio by dividing both numerator and denominator by their greatest common divisor.

class (Num a, Ord a) => Real a where

Real numbers.

The Haskell report defines no laws for Real, however Real instances are customarily expected to adhere to the following law:

Coherence with fromRational

if the type also implements Fractional, then fromRational is a left inverse for toRational, i.e. fromRational (toRational i) = i

Methods

toRational :: a -> Rational

the rational equivalent of its real argument with full precision

class (Real a, Enum a) => Integral a where

Integral numbers, supporting integer division.

The Haskell Report defines no laws for Integral. However, Integral instances are customarily expected to define a Euclidean domain and have the following properties for the div/mod and quot/rem pairs, given suitable Euclidean functions f and g:

x = y * quot x y + rem x y with rem x y = fromInteger 0 or g (rem x y) < g y

x = y * div x y + mod x y with mod x y = fromInteger 0 or f (mod x y) < f y

An example of a suitable Euclidean function, for Integer's instance, is abs.

In addition, toInteger should be total, and fromInteger should be a left inverse for it, i.e. fromInteger (toInteger i) = i.

Minimal complete definition

quotRem, toInteger

Methods

quot :: a -> a -> a infixl 7

integer division truncated toward zero

WARNING: This function is partial (because it throws when 0 is passed as the divisor) for all the integer types in base.

rem :: a -> a -> a infixl 7

integer remainder, satisfying

(x `quot` y)*y + (x `rem` y) == x

WARNING: This function is partial (because it throws when 0 is passed as the divisor) for all the integer types in base.

div :: a -> a -> a infixl 7

integer division truncated toward negative infinity

WARNING: This function is partial (because it throws when 0 is passed as the divisor) for all the integer types in base.

mod :: a -> a -> a infixl 7

integer modulus, satisfying

(x `div` y)*y + (x `mod` y) == x

WARNING: This function is partial (because it throws when 0 is passed as the divisor) for all the integer types in base.

quotRem :: a -> a -> (a, a)

simultaneous quot and rem

WARNING: This function is partial (because it throws when 0 is passed as the divisor) for all the integer types in base.

divMod :: a -> a -> (a, a)

simultaneous div and mod

WARNING: This function is partial (because it throws when 0 is passed as the divisor) for all the integer types in base.

toInteger :: a -> Integer

conversion to Integer

class Num a => Fractional a where

Fractional numbers, supporting real division.

The Haskell Report defines no laws for Fractional. However, (+) and (*) are customarily expected to define a division ring and have the following properties:

recip gives the multiplicative inverse

x * recip x = recip x * x = fromInteger 1

Totality of toRational

toRational is total

Coherence with toRational

if the type also implements Real, then fromRational is a left inverse for toRational, i.e. fromRational (toRational i) = i

Note that it isn't customarily expected that a type instance of Fractional implement a field. However, all instances in base do.

Minimal complete definition

fromRational, (recip | (/))

Methods

(/) :: a -> a -> a infixl 7

Fractional division.

recip :: a -> a

Reciprocal fraction.

fromRational :: Rational -> a

Conversion from a Rational (that is Ratio Integer). A floating literal stands for an application of fromRational to a value of type Rational, so such literals have type (Fractional a) => a.

class (Real a, Fractional a) => RealFrac a where

Extracting components of fractions.

Minimal complete definition

properFraction

Methods

properFraction :: Integral b => a -> (b, a)

The function properFraction takes a real fractional number x and returns a pair (n,f) such that x = n+f, and:

n is an integral number with the same sign as x; and

f is a fraction with the same type and sign as x, and with absolute value less than 1.

The default definitions of the ceiling, floor, truncate and round functions are in terms of properFraction.

truncate :: Integral b => a -> b

truncate x returns the integer nearest x between zero and x

round :: Integral b => a -> b

round x returns the nearest integer to x; the even integer if x is equidistant between two integers

ceiling :: Integral b => a -> b

ceiling x returns the least integer not less than x

floor :: Integral b => a -> b

floor x returns the greatest integer not greater than x

numericEnumFrom :: Fractional a => a -> [a]

numericEnumFromThen :: Fractional a => a -> a -> [a]

numericEnumFromTo :: (Ord a, Fractional a) => a -> a -> [a]

numericEnumFromThenTo :: (Ord a, Fractional a) => a -> a -> a -> [a]

fromIntegral :: (Integral a, Num b) => a -> b

General coercion from Integral types.

WARNING: This function performs silent truncation if the result type is not at least as big as the argument's type.

realToFrac :: (Real a, Fractional b) => a -> b

General coercion to Fractional types.

WARNING: This function goes through the Rational type, which does not have values for NaN for example. This means it does not round-trip.

For Double it also behaves differently with or without -O0:

Prelude> realToFrac nan -- With -O0
-Infinity
Prelude> realToFrac nan
NaN

showSigned

Arguments

:: Real a

=> (a -> ShowS)
a function that can show unsigned values

-> Int
the precedence of the enclosing context

-> a
the value to show

-> ShowS

Converts a possibly-negative Real value to a string.

even :: Integral a => a -> Bool

odd :: Integral a => a -> Bool

(^) :: (Num a, Integral b) => a -> b -> a infixr 8

raise a number to a non-negative integral power

powImpl :: (Num a, Integral b) => a -> b -> a

powImplAcc :: (Num a, Integral b) => a -> b -> a -> a

(^^) :: (Fractional a, Integral b) => a -> b -> a infixr 8

raise a number to an integral power

(^%^) :: Integral a => Rational -> a -> Rational

(^^%^^) :: Integral a => Rational -> a -> Rational

gcd :: Integral a => a -> a -> a

gcd x y is the non-negative factor of both x and y of which every common factor of x and y is also a factor; for example gcd 4 2 = 2, gcd (-4) 6 = 2, gcd 0 4 = 4. gcd 0 0 = 0. (That is, the common divisor that is "greatest" in the divisibility preordering.)

Note: Since for signed fixed-width integer types, abs minBound < 0, the result may be negative if one of the arguments is minBound (and necessarily is if the other is 0 or minBound) for such types.

lcm :: Integral a => a -> a -> a

lcm x y is the smallest positive integer that both x and y divide.

integralEnumFrom :: (Integral a, Bounded a) => a -> [a]

integralEnumFromThen :: (Integral a, Bounded a) => a -> a -> [a]

integralEnumFromTo :: Integral a => a -> a -> [a]

integralEnumFromThenTo :: Integral a => a -> a -> a -> [a]

data FractionalExponentBase

Constructors

Base2

Base10

mkRationalBase2 :: Rational -> Integer -> Rational

mkRationalBase10 :: Rational -> Integer -> Rational

mkRationalWithExponentBase :: Rational -> Integer -> FractionalExponentBase -> Rational