GHC.Records

Description

This module defines the HasField class used by the OverloadedRecordFields extension. See the <https://ghc.haskell.org/trac/ghc/wiki/Records/OverloadedRecordFields wiki page> for more details.

Documentation

class HasField (x :: k) r a | x r -> a where

Constraint representing the fact that the field x belongs to the record type r and has field type a. This will be solved automatically, but manual instances may be provided as well.

Methods

getField :: r -> a

Selector function to extract the field from the record.