GHC.Stack

Description

Access to GHC's call-stack simulation

Since: 4.5.0.0

Call stack

currentCallStack :: IO [String]

returns a '[String]' representing the current call stack. This can be useful for debugging.

The implementation uses the call-stack simulation maintined by the profiler, so it only works if the program was compiled with -prof and contains suitable SCC annotations (e.g. by using -fprof-auto). Otherwise, the list returned is likely to be empty or uninformative.

Since: 4.5.0.0

whoCreated :: a -> IO [String]

Get the stack trace attached to an object.

Since: 4.5.0.0

errorWithStackTrace :: String -> a

Like the function error, but appends a stack trace to the error message if one is available.

Since: 4.7.0.0

Internals

data CostCentreStack

data CostCentre

getCurrentCCS :: dummy -> IO (Ptr CostCentreStack)

getCCSOf :: a -> IO (Ptr CostCentreStack)

ccsCC :: Ptr CostCentreStack -> IO (Ptr CostCentre)

ccsParent :: Ptr CostCentreStack -> IO (Ptr CostCentreStack)

ccLabel :: Ptr CostCentre -> IO CString

ccModule :: Ptr CostCentre -> IO CString

ccSrcSpan :: Ptr CostCentre -> IO CString

ccsToStrings :: Ptr CostCentreStack -> IO [String]

renderStack :: [String] -> String