GHC.TypeLits

Description

This module is an internal GHC module. It declares the constants used in the implementation of type-level natural numbers. The programmer interface for working with type-level naturals should be defined in a separate library.

Since: 4.6.0.0

Kinds

data Nat

(Kind) This is the kind of type-level natural numbers.

data Symbol

(Kind) This is the kind of type-level symbols.

Linking type and value level

class KnownNat n

This class gives the integer associated with a type-level natural. There are instances of the class for every concrete literal: 0, 1, 2, etc.

Since: 4.7.0.0

Minimal complete definition

natSing

natVal :: forall n proxy. KnownNat n => proxy n -> Integer

Since: 4.7.0.0

natVal' :: forall n. KnownNat n => Proxy# n -> Integer

Since: 4.8.0.0

class KnownSymbol n

This class gives the string associated with a type-level symbol. There are instances of the class for every concrete literal: "hello", etc.

Since: 4.7.0.0

Minimal complete definition

symbolSing

symbolVal :: forall n proxy. KnownSymbol n => proxy n -> String

Since: 4.7.0.0

symbolVal' :: forall n. KnownSymbol n => Proxy# n -> String

Since: 4.8.0.0

data SomeNat

This type represents unknown type-level natural numbers.

Constructors

forall n . KnownNat n => SomeNat (Proxy n)
Since: 4.7.0.0

data SomeSymbol

This type represents unknown type-level symbols.

Constructors

forall n . KnownSymbol n => SomeSymbol (Proxy n)
Since: 4.7.0.0

someNatVal :: Integer -> Maybe SomeNat

Convert an integer into an unknown type-level natural.

Since: 4.7.0.0

someSymbolVal :: String -> SomeSymbol

Convert a string into an unknown type-level symbol.

Since: 4.7.0.0

sameNat :: (KnownNat a, KnownNat b) => Proxy a -> Proxy b -> Maybe (a :~: b)

We either get evidence that this function was instantiated with the same type-level numbers, or Nothing.

Since: 4.7.0.0

sameSymbol :: (KnownSymbol a, KnownSymbol b) => Proxy a -> Proxy b -> Maybe (a :~: b)

We either get evidence that this function was instantiated with the same type-level symbols, or Nothing.

Since: 4.7.0.0

Functions on type literals

type (<=) x y = (x <=? y) ~ True infix 4

Comparison of type-level naturals, as a constraint.

type family m <=? n :: Bool infix 4

Comparison of type-level naturals, as a function. NOTE: The functionality for this function should be subsumed by CmpNat, so this might go away in the future. Please let us know, if you encounter discrepancies between the two.

type family m + n :: Nat infixl 6

Addition of type-level naturals.

type family m * n :: Nat infixl 7

Multiplication of type-level naturals.

type family m ^ n :: Nat infixr 8

Exponentiation of type-level naturals.

type family m - n :: Nat infixl 6

Subtraction of type-level naturals.

Since: 4.7.0.0

type family CmpNat m n :: Ordering

Comparison of type-level naturals, as a function.

Since: 4.7.0.0

type family CmpSymbol m n :: Ordering

Comparison of type-level symbols, as a function.

Since: 4.7.0.0