GHC.TypeLits.Internal

Description

This module exports the Type Literal kinds as well as the comparison type families for those kinds. It is needed to prevent module cycles while still allowing these identifiers to be imported in Ord.

Since: base-4.16.0.0

Documentation

data Symbol

(Kind) This is the kind of type-level symbols.

type family CmpSymbol (m :: Symbol) (n :: Symbol) :: Ordering

Comparison of type-level symbols, as a function.

Since: base-4.7.0.0

type family CmpChar (a :: Char) (b :: Char) :: Ordering

Comparison of type-level characters.

Since: base-4.16.0.0