System.Info

Description

Information about the characteristics of the host system lucky enough to run your program.

Documentation

os :: String

The operating system on which the program is running.

arch :: String

The machine architecture on which the program is running.

compilerName :: String

The Haskell implementation with which the program was compiled or is being interpreted.

compilerVersion :: Version

The version of compilerName with which the program was compiled or is being interpreted.