System.Posix.Types

Description

POSIX data types: Haskell equivalents of the types defined by the <sys/types.h> C header on a POSIX system.

POSIX data types

Platform differences

This module contains platform specific information about types. __/As such the types presented on this page reflect the platform on which the documentation was generated and may not coincide with the types on your platform./__

newtype CDev

Constructors

CDev Word64

newtype CIno

Constructors

CIno Word64

newtype CMode

Constructors

CMode Word32

newtype COff

Constructors

COff Int64

newtype CPid

Constructors

CPid Int32

newtype CSsize

Constructors

CSsize Int64

newtype CGid

Constructors

CGid Word32

newtype CNlink

Constructors

CNlink Word64

newtype CUid

Constructors

CUid Word32

newtype CCc

Constructors

CCc Word8

newtype CSpeed

Constructors

CSpeed Word32

newtype CTcflag

Constructors

CTcflag Word32

newtype CRLim

Constructors

CRLim Word64

newtype CBlkSize

Since: base-4.10.0.0

Constructors

CBlkSize Int64

newtype CBlkCnt

Since: base-4.10.0.0

Constructors

CBlkCnt Int64

newtype CClockId

Since: base-4.10.0.0

Constructors

CClockId Int32

newtype CFsBlkCnt

Since: base-4.10.0.0

Constructors

CFsBlkCnt Word64

newtype CFsFilCnt

Since: base-4.10.0.0

Constructors

CFsFilCnt Word64

newtype CId

Since: base-4.10.0.0

Constructors

CId Word32

newtype CKey

Since: base-4.10.0.0

Constructors

CKey Int32

newtype CTimer

Since: base-4.10.0.0

Constructors

CTimer (Ptr ())

newtype Fd

Constructors

Fd CInt

type LinkCount = CNlink

type UserID = CUid

type GroupID = CGid

type ByteCount = CSize

type ClockTick = CClock

type EpochTime = CTime

type FileOffset = COff

type ProcessID = CPid

type ProcessGroupID = CPid

type DeviceID = CDev

type FileID = CIno

type FileMode = CMode

type Limit = CLong