System.Posix.Types

Description

POSIX data types: Haskell equivalents of the types defined by the <sys/types.h> C header on a POSIX system.

POSIX data types

Platform differences

This module contains platform specific information about types. __/As such the types presented on this page reflect the platform on which the documentation was generated and may not coincide with the types on your platform./__

newtype CDev

Constructors

CDev Int32

newtype CIno

Constructors

CIno Word64

newtype CMode

Constructors

CMode Word16

newtype COff

Constructors

COff Int64

newtype CPid

Constructors

CPid Int32

newtype CSsize

Constructors

CSsize Int64

newtype CGid

Constructors

CGid Word32

newtype CNlink

Constructors

CNlink Word16

newtype CUid

Constructors

CUid Word32

newtype CCc

Constructors

CCc Word8

newtype CSpeed

Constructors

CSpeed Word64

newtype CTcflag

Constructors

CTcflag Word64

newtype CRLim

Constructors

CRLim Word64

newtype CBlkSize

Since: 4.10.0.0

Constructors

CBlkSize Int32

newtype CBlkCnt

Since: 4.10.0.0

Constructors

CBlkCnt Int64

newtype CClockId

Since: 4.10.0.0

Constructors

CClockId Word32

newtype CFsBlkCnt

Since: 4.10.0.0

Constructors

CFsBlkCnt Word32

newtype CFsFilCnt

Since: 4.10.0.0

Constructors

CFsFilCnt Word32

newtype CId

Since: 4.10.0.0

Constructors

CId Word32

newtype CKey

Since: 4.10.0.0

Constructors

CKey Int32

newtype Fd

Constructors

Fd CInt

type LinkCount = CNlink

type UserID = CUid

type GroupID = CGid

type ByteCount = CSize

type ClockTick = CClock

type EpochTime = CTime

type FileOffset = COff

type ProcessID = CPid

type ProcessGroupID = CPid

type DeviceID = CDev

type FileID = CIno

type FileMode = CMode

type Limit = CLong