System.Posix.Types

Description

POSIX data types: Haskell equivalents of the types defined by the <sys/types.h> C header on a POSIX system.

POSIX data types

Platform differences

This module contains platform specific information about types. __/As such the types presented on this page reflect the platform on which the documentation was generated and may not coincide with the types on your platform./__

newtype CDev

Constructors

CDev Word32

newtype CIno

Constructors

CIno Word16

newtype CMode

Constructors

CMode Word16

newtype COff

Constructors

COff Int64

newtype CPid

Constructors

CPid Int64

newtype CSsize

Constructors

CSsize Int64

newtype CClockId

Since: base-4.10.0.0

Constructors

CClockId Int32

newtype Fd

Constructors

Fd CInt

type ByteCount = CSize

type ClockTick = CClock

type EpochTime = CTime

type FileOffset = COff

type ProcessID = CPid

type ProcessGroupID = CPid

type DeviceID = CDev

type FileID = CIno

type FileMode = CMode

type Limit = CLong